### 2025-04-18 c538361
docs: fix DOIs in the paper
