+++ paper/paper.bib
@@ -17,6 +17,7 @@
     journal={Journal of Machine Learning Research},
     volume={12},
     pages={2825--2830},
+    doi={10.5555/1953048.2078195},
     url={http://jmlr.org/papers/v12/pedregosa11a.html},
     year={2011}
 }
@@ -26,6 +27,7 @@
     year={2017},
     title={The {UCI} Machine Learning Repository},
     url={https://archive.ics.uci.edu},
+    doi={10.17616/R3T91Q},
     institution={University of California, Irvine, School of Information and Computer Sciences}
 }
 
@@ -37,7 +39,7 @@
     number={3},
     pages={421--441},
     year={2018},
-    doi={https://doi.org/10.1109/TEVC.2018.2868770},
+    doi={10.1109/TEVC.2018.2868770},
     publisher={IEEE}
 }
 
@@ -49,7 +51,7 @@
     number={5},
     pages={882--895},
     year={2020},
-    doi={https://doi.org/10.1109/TEVC.2020.2968743},
+    doi={10.1109/TEVC.2020.2968743},
     publisher={IEEE}
 }
 
@@ -60,7 +62,7 @@
     volume={8},
     pages={150113--150129},
     year={2020},
-    doi={https://doi.org/10.1109/ACCESS.2020.3016679},
+    doi={10.1109/ACCESS.2020.3016679},
     publisher={IEEE}
 }
 
@@ -72,7 +74,7 @@
     number={1},
     pages={1--42},
     year={2020},
-    doi={https://doi.org/10.1186/s40537-020-00381-y},
+    doi={10.1186/s40537-020-00381-y},
     publisher={SpringerOpen}
 }
 
@@ -82,7 +84,7 @@
     journal={The Journal of Supercomputing},
     pages={1--30},
     year={2023},
-    doi={https://doi.org/10.1007/s11227-023-05226-y},
+    doi={10.1007/s11227-023-05226-y},
     publisher={Springer}
 }
 
@@ -94,7 +96,7 @@
     number={11},
     pages={4501},
     year={2024},
-    doi={https://doi.org/10.3390/app14114501},
+    doi={10.3390/app14114501},
     publisher={Multidisciplinary Digital Publishing Institute}
 }
 
@@ -106,7 +108,7 @@
     number={3},
     pages={1575--1637},
     year={2024},
-    doi={https://doi.org/10.1007/s10115-023-02010-5},
+    doi={10.1007/s10115-023-02010-5},
     publisher={Springer}
 }
 
@@ -118,5 +120,5 @@
     pages={113327},
     year={2025},
     issn={0950-7051},
-    doi={https://doi.org/10.1016/j.knosys.2025.113327},
+    doi={10.1016/j.knosys.2025.113327},
 }
