### 2025-04-18 26c5329
docs: fix DOI in scikit-learn reference
