+++ paper/paper.bib
@@ -17,8 +17,7 @@
     journal={Journal of Machine Learning Research},
     volume={12},
     pages={2825--2830},
-    doi={10.5555/1953048.2078195},
-    url={http://jmlr.org/papers/v12/pedregosa11a.html},
+    url={https://dl.acm.org/doi/10.5555/1953048.2078195},
     year={2011}
 }
 
