### 2025-04-18 212af81
feat: add paper status badge
