### 2025-04-19 cfb6b9d
refactor: define dataset config file to enable support for regression and clustering tasks
