### 2025-04-19 1bb1400
chore: reduce max generations for short optimization runs
