### 2025-04-19 d95d535
feat: add Infrared Thermography Temperature regression datasets
