### 2025-04-19 3c4f825
feat: log number of rows removed due to NaN values during data cleaning
