### 2025-04-19 6a35855
feat: add regression metrics class
