+++ tests/test_metrics.py
@@ -1,10 +1,10 @@
 import pytest
 import numpy as np
-from pyccea.utils.metrics import ClassificationMetrics
+from pyccea.utils.metrics import ClassificationMetrics, RegressionMetrics
 
 
-class DummyClassifier:
-    """A simple mock classifier returning pre-defined predictions."""
+class DummyEstimator:
+    """A simple mock estimator returning pre-defined predictions."""
     def __init__(self, predictions):
         self._predictions = predictions
 
@@ -18,7 +18,7 @@ def test_classification_metrics_binary() -> None:
     y_test = np.array([0, 1, 0, 1])
     y_pred = np.array([0, 1, 0, 0])
 
-    estimator = DummyClassifier(predictions=y_pred)
+    estimator = DummyEstimator(predictions=y_pred)
 
     metrics = ClassificationMetrics(n_classes=2)
     metrics.compute(estimator, X_test, y_test)
@@ -32,8 +32,8 @@ def test_classification_metrics_binary() -> None:
         "specificity": 1.0
     }
 
-    for key, expected in expected_values.items():
-        assert metrics.values[key] == pytest.approx(expected, abs=1e-4)
+    for metric_name, expected in expected_values.items():
+        assert metrics.values[metric_name] == pytest.approx(expected, abs=1e-4), f"{metric_name} mismatch"
 
 
 def test_classification_metrics_multiclass() -> None:
@@ -42,7 +42,7 @@ def test_classification_metrics_multiclass() -> None:
     y_test = np.array([0, 1, 2, 0, 1, 2])
     y_pred = np.array([0, 2, 1, 0, 1, 2])
 
-    estimator = DummyClassifier(predictions=y_pred)
+    estimator = DummyEstimator(predictions=y_pred)
 
     metrics = ClassificationMetrics(n_classes=3)
     metrics.compute(estimator, X_test, y_test)
@@ -59,18 +59,27 @@ def test_classification_metrics_multiclass() -> None:
         assert metrics.values[metric_name] == pytest.approx(expected, abs=1e-4), f"{metric_name} mismatch"
 
 
-def test_classification_metrics_invalid_input() -> None:
-    """Test invalid input for classification metrics."""
-    X_test = np.array([[0], [1]])
-    y_test = np.array([0, 1])
-    # Wrong length on purpose
-    y_pred = np.array([0, 1, 0])
+def test_regression_metrics() -> None:
+    """Test regression metrics."""
+    X_test = np.array([[0], [1], [2], [3], [4]])
+    y_test = np.array([3.0, -0.5, 2.0, 7.0, 4.2])
+    y_pred = np.array([2.5, 0.0, 2.1, 7.8, 5.3])
 
-    estimator = DummyClassifier(predictions=y_pred)
-    metrics = ClassificationMetrics(n_classes=2)
+    estimator = DummyEstimator(predictions=y_pred)
 
-    with pytest.raises(ValueError):
-        metrics.compute(estimator, X_test, y_test)
+    metrics = RegressionMetrics()
+    metrics.compute(estimator, X_test, y_test)
+
+    expected_values = {
+        "mae": 0.6,
+        "mse": 0.472,
+        "rmse": 0.687,
+        "r2": 0.9229,
+        "mape": 31.8571
+    }
+
+    for metric_name, expected in expected_values.items():
+        assert metrics.values[metric_name] == pytest.approx(expected, abs=1e-4), f"{metric_name} mismatch"
 
 
 def test_classification_metrics_verbose_output(caplog) -> None:
@@ -79,11 +88,32 @@ def test_classification_metrics_verbose_output(caplog) -> None:
     y_test = np.array([0, 1])
     y_pred = np.array([0, 1])
 
[…]
