### 2025-04-19 b0a38dd
test: improve unit test coverage for DataLoader
