+++ tests/test_dataloader.py
@@ -1,5 +1,9 @@
+import os
 import pytest
 import logging
+import tempfile
+import numpy as np
+import pandas as pd
 from pyccea.utils.datasets import DataLoader
 
 
@@ -88,3 +92,204 @@ def test_load_data_invalid_dataset_name(complete_data_conf: dict) -> None:
     with pytest.raises(ValueError):
         dl = DataLoader("invalid_dataset", complete_data_conf)
         dl._load()
+
+
+def test_check_header_true(complete_data_conf: dict) -> None:
+    """Test _check_header returns True when CSV has a header."""
+    csv_content = "col1,1,col3\n1,2,3\n"
+    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".csv") as tmp:
+        tmp.write(csv_content)
+        tmp.flush()
+        tmp_path = tmp.name
+
+    loader = DataLoader("dummy", complete_data_conf)
+    assert loader._check_header(tmp_path) is True
+
+    os.remove(tmp_path)
+
+
+def test_check_header_false(complete_data_conf: dict) -> None:
+    """Test _check_header returns False when CSV has no header."""
+    csv_content = "1,2,3\n4,5,6\n"
+    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".csv") as tmp:
+        tmp.write(csv_content)
+        tmp.flush()
+        tmp_path = tmp.name
+
+    loader = DataLoader("dummy", complete_data_conf)
+    logging.warning(loader._check_header(tmp_path))
+    assert loader._check_header(tmp_path) is False
+
+    os.remove(tmp_path)
+
+
+def test_get_input_and_output(complete_data_conf: dict) -> None:
+    """Test _get_input and _get_output extract correct columns."""
+    loader = DataLoader("dummy", complete_data_conf)
+    loader.data = pd.DataFrame({
+        0: [1, 2],
+        1: [3, 4],
+        2: [5, 6],
+        3: ["train", "test"]
+    })
+
+    X = loader._get_input()
+    y = loader._get_output()
+
+    expected_X = pd.DataFrame({0: [1, 2], 1: [3, 4]})
+    expected_y = pd.Series([5, 6], name=2)
+
+    pd.testing.assert_frame_equal(X, expected_X)
+    pd.testing.assert_series_equal(y, expected_y)
+
+
+def test_preprocess(monkeypatch, complete_data_conf: dict) -> None:
+    """Test _preprocess replaces ? with NaN, drops rows if required."""
+    loader = DataLoader("dummy", complete_data_conf)
+    loader.data = pd.DataFrame({
+        0: [1, "?", 3],
+        1: [4, 5, None],
+        2: [0, 1, 0],
+        3: ["train", "test", "train"]
+    })
+
+    monkeypatch.setitem(DataLoader.DATASETS, "dummy", {"task": "classification"})
+
+    loader._preprocess()
+
+    # Check NaN replaced and rows dropped
+    assert not loader.data.isin(["?"]).any().any()
+    assert loader.n_examples == 1
+    assert loader.n_features == 2
+    assert loader.n_classes == 1
+    assert loader.classes == [0]
+
+
+def test_split_preset(monkeypatch, complete_data_conf: dict) -> None:
+    """Test _split with preset splits."""
+    loader = DataLoader("dummy", complete_data_conf)
+    loader.data = pd.DataFrame({
+        0: [1, 2, 3, 4, 5],
+        1: [6, 7, 8, 9, 0],
+        2: [0, 1, 0, 1, 1],
+        3: ["train", "train", "train", "test", "test"]
+    })
+    loader.X = loader._get_input()
+    loader.y = loader._get_output()
+
+    loader._split()
[…]
