### 2025-04-19 7cf3860
docs: improve snippet code
