+++ tests/test_dataloader.py
@@ -4,7 +4,9 @@ import logging
 import tempfile
 import numpy as np
 import pandas as pd
+from unittest.mock import patch
 from pyccea.utils.datasets import DataLoader
+from sklearn.model_selection import KFold, LeaveOneOut, StratifiedKFold
 
 
 # Disable logging output during tests
@@ -47,6 +49,14 @@ def test_parse_splitter_parameters_missing_kfolds(complete_data_conf: dict) -> N
         DataLoader("dummy_dataset", complete_data_conf)
 
 
+def test_parse_splitter_parameters_preset_true_and_test_ratio(complete_data_conf: dict) -> None:
+    """Test DataLoader initialization with preset and test ratio."""
+    complete_data_conf["splitter"]["preset"] = True
+    complete_data_conf["splitter"]["test_ratio"] = 0.2
+    dl = DataLoader("dummy_dataset", complete_data_conf)
+    assert dl.test_ratio is None
+
+
 def test_parse_splitter_parameters_missing_test_ratio_and_preset_false(complete_data_conf: dict) -> None:
     """Test DataLoader initialization with missing test ratio when preset is false."""
     complete_data_conf["splitter"]["preset"] = False
@@ -87,11 +97,141 @@ def test_parse_normalization_parameters_method_and_normalize_false(complete_data
         DataLoader("dummy_dataset", complete_data_conf)
 
 
+def test_get_ready_calls_all_methods(mocker, complete_data_conf: dict) -> None:
+    """Test get_ready method when normalize is True."""
+    loader = DataLoader("dummy", complete_data_conf)
+
+    # Mock _load to assign synthetic data
+    def _mock_load():
+        loader.data = pd.DataFrame({
+            0: [1, 2, 3, 4, 5],
+            1: [6, 7, 8, 9, 0],
+            2: [0, 1, 0, 1, 1],
+            3: ["train", "train", "train", "test", "test"]
+        })
+
+    mocker.patch.object(loader, "_load", side_effect=_mock_load)
+
+    # Mock the other methods just to track their calls
+    mocker.patch.object(loader, "_preprocess")
+    mocker.patch.object(loader, "_split")
+    mocker.patch.object(loader, "_normalize_subsets")
+    mocker.patch.object(loader, "_model_selection")
+
+    # Act
+    loader.get_ready()
+
+    # Assert: all methods called once
+    loader._load.assert_called_once()
+    loader._preprocess.assert_called_once()
+    loader._split.assert_called_once()
+    loader._normalize_subsets.assert_called_once()
+    loader._model_selection.assert_called_once()
+
+
+def test_get_ready_without_normalize(mocker, complete_data_conf: dict) -> None:
+    """Test get_ready method when normalize is False."""
+    complete_data_conf["normalization"]["normalize"] = False
+    del complete_data_conf["normalization"]["method"]
+    loader = DataLoader("dummy", complete_data_conf)
+
+    # Mock _load to assign synthetic data
+    def _mock_load():
+        loader.data = pd.DataFrame({
+            0: [1, 2, 3, 4, 5],
+            1: [6, 7, 8, 9, 0],
+            2: [0, 1, 0, 1, 1],
+            3: ["train", "train", "train", "test", "test"]
+        })
+
+    mocker.patch.object(loader, "_load", side_effect=_mock_load)
+
+    # Mock the other methods just to track their calls
+    mocker.patch.object(loader, "_preprocess")
+    mocker.patch.object(loader, "_split")
+    mocker.patch.object(loader, "_normalize_subsets")
+    mocker.patch.object(loader, "_model_selection")
+
+    # Act
+    loader.get_ready()
+
[…]
