+++ requirements.txt
@@ -10,3 +10,4 @@ pytest==8.3.5
 pytest-cov==6.1.1
 pytest-asyncio==0.26.0
 pytest-deadfixtures==2.2.1
+pytest-mock==3.14.0
