### 2025-04-19 fadfe6c
chore: change package setup
