### 2025-04-19 659daea
feat: add RegressionModel class
