+++ pyccea/utils/models.py
@@ -1,25 +1,201 @@
 import logging
 import warnings
 import numpy as np
-from sklearn.svm import SVC
+from sklearn.svm import SVC, SVR
 from sklearn.neighbors import KNeighborsClassifier
-from sklearn.ensemble import RandomForestClassifier
-from sklearn.linear_model import LogisticRegression
 from sklearn.exceptions import UndefinedMetricWarning
 from sklearn.model_selection import RandomizedSearchCV
 from sklearn.naive_bayes import ComplementNB, GaussianNB, MultinomialNB
+from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
+from sklearn.linear_model import LogisticRegression, LinearRegression, Ridge, Lasso, ElasticNet
 
 
-class ClassificationModel():
-    """
-    Load a machine learning model, adjust its hyperparameters and get the best model.
+class RegressionModel():
+    """Load a regression model, adjust its hyperparameters and get the best model.
 
     Attributes
     ----------
-    estimator: sklearn model object
+    estimator : sklearn model object
         Trained model. In case optimize is True, it is the model that was chosen by the Randomized
         Search, i.e., estimator which gave the best result on the validation data.
-    hyperparams: dict
+    hyperparams : dict
+        Hyperparameters of the model. In case optimize is True, it is the best hyperparameters
+        used to fit the machine learning model.
+    """
+
+    models = {
+        "linear": (LinearRegression, {}),
+        "ridge": (Ridge, {}),
+        "lasso": (Lasso, {}),
+        "elastic_net": (ElasticNet, {}),
+        "random_forest": (RandomForestRegressor, {}),
+        "support_vector_machine": (SVR, {}),
+    }
+
+    def __init__(self, model_type: str):
+        """Initialize the regression model.
+
+        Parameters
+        ----------
+        model_type : str
+            Name of the machine learning model that will be fitted for a regression task.
+        """
+        # Check if the chosen regression model is available
+        if model_type not in RegressionModel.models.keys():
+            raise AssertionError(
+                f"Model type '{model_type}' was not found. "
+                f"The available models are {', '.join(RegressionModel.models.keys())}."
+            )
+        # Initialize regression model
+        self.model_type = model_type
+        model_class, params = RegressionModel.models[self.model_type]
+        self.estimator = model_class(**params)
+        # Initialize logger with info level
+        logging.basicConfig(encoding="utf-8", level=logging.INFO)
+        # Suppress warnings
+        warnings.filterwarnings(action="ignore", category=FutureWarning)
+
+
+    def _model_selection(self, 
+                         X_train: np.ndarray,
+                         y_train: np.ndarray,
+                         n_iter: int = 100,
+                         seed: int = None,
+                         kfolds: int = 5):
+        """Optimize the hyperparameters of the model and return the best model found.
+
+        Parameters
+        ----------
+        X_train: np.ndarray
+            Train input data.
+        y_train: np.ndarray
+            Train output data.
+        n_iter: int, default 100
+            Number of hyperparameter settings that are sampled. It trades off runtime and quality
+            of the solution.
+        seed: int, default None
+            Controls the shuffling applied for subsampling the data.
+        kfolds: int, default 5
+            Number of folds in the k-fold cross validation.
+
+        Returns
+        -------
+        estimator: sklearn model object
+            Trained model. It is the model that was chosen by the Randomized Search, i.e., estimator
+            which gave the best result on the validation data.
+        hyperparams: dict
+            Hyperparameters of the model. It is the best hyperparameters used to fit the machine
+            learning model.
+        """
+        # Grid of possible hyperparameter values
+        if self.model_type == "ridge":
[…]
