+++ tests/test_models.py
@@ -1,61 +1,81 @@
 import pytest
-from sklearn.base import ClassifierMixin
+from sklearn.base import ClassifierMixin, RegressorMixin
 
-from pyccea.utils.models import ClassificationModel
+from pyccea.utils.models import ClassificationModel, RegressionModel
 
 
-def test_model_initialization_valid() -> None:
-    """Test model instantiation for each available model."""
+def test_classification_model_initialization_valid() -> None:
+    """Test model instantiation for each available classification model."""
     for model_type in ClassificationModel.models.keys():
         model = ClassificationModel(model_type)
         assert isinstance(model.estimator, ClassifierMixin)
 
 
-def test_model_initialization_invalid() -> None:
-    """Test invalid model type raises AssertionError."""
+def test_regression_model_initialization_valid() -> None:
+    """Test model instantiation for each available regression model."""
+    for model_type in RegressionModel.models.keys():
+        model = RegressionModel(model_type)
+        assert isinstance(model.estimator, RegressorMixin)
+
+
+def test_classification_model_initialization_invalid() -> None:
+    """Test invalid classification model type raises AssertionError."""
     with pytest.raises(AssertionError):
-        ClassificationModel("invalid_model_name")
+        ClassificationModel("linear_regression")
 
 
-def test_train_without_optimization(classification_data) -> None:
-    """Test model training without hyperparameter optimization."""
+def test_regression_model_initialization_invalid() -> None:
+    """Test invalid regression model type raises AssertionError."""
+    with pytest.raises(AssertionError):
+        RegressionModel("logistic_regression")
+
+
+def test_train_classifier_without_optimization(classification_data) -> None:
+    """Test classification model training without hyperparameter optimization."""
     X_train, y_train = classification_data
     model = ClassificationModel("random_forest")
     model.train(X_train, y_train, optimize=False)
     
     # Check that the estimator is fitted and hyperparams available
     assert hasattr(model.estimator, "predict")
+    assert isinstance(model.estimator, ClassifierMixin)
     assert isinstance(model.hyperparams, dict)
-    assert "n_estimators" in model.hyperparams  # for RandomForestClassifier
 
 
-def test_train_with_optimization(classification_data) -> None:
-    """Test model training with hyperparameter optimization."""
+def test_train_regressor_without_optimization(regression_data) -> None:
+    """Test regression model training without hyperparameter optimization."""
+    X_train, y_train = regression_data
+    model = RegressionModel("random_forest")
+    model.train(X_train, y_train, optimize=False)
+    
+    # Check that the estimator is fitted and hyperparams available
+    assert hasattr(model.estimator, "predict")
+    assert isinstance(model.estimator, RegressorMixin)
+    assert isinstance(model.hyperparams, dict)
+
+
+def test_train_classifier_with_optimization(classification_data) -> None:
+    """Test classification model training with hyperparameter optimization."""
     X_train, y_train = classification_data
     model = ClassificationModel("random_forest")
     model.train(X_train, y_train, optimize=True, n_iter=5, kfolds=3, seed=42)
     
     # Check estimator is fitted and hyperparams are available after optimization
     assert hasattr(model.estimator, "predict")
+    assert isinstance(model.estimator, ClassifierMixin)
     assert isinstance(model.hyperparams, dict)
-    assert "n_estimators" in model.hyperparams
 
 
-def test_model_selection_returns_expected(classification_data) -> None:
-    """Test _model_selection returns estimator and hyperparams."""
[…]
