+++ tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from sklearn.datasets import make_classification
+from sklearn.datasets import make_classification, make_regression
 
 
 @pytest.fixture
@@ -20,6 +20,19 @@ def classification_data():
         n_features=5,
         n_classes=2,
         n_informative=3,
+        random_state=42
+    )
+    return X, y
+
+
+@pytest.fixture
+def regression_data():
+    """Fixture for simple regression dataset."""
+    X, y = make_regression(
+        n_samples=100,
+        n_features=5,
+        n_informative=3,
+        noise=5.0,
         random_state=42
     )
     return X, y
