### 2025-04-27 eb1e302
feat: extend CCEAs to support regression tasks
