+++ pyccea/coevolution/ccga.py
@@ -31,7 +31,7 @@ class CCGA(CCEA):
             model_type=self.conf["wrapper"]["model_type"],
             eval_function=self.conf["evaluation"]["eval_function"],
             eval_mode=self.eval_mode,
-            n_classes=self.data.n_classes
+            n_classes=getattr(self.data, "n_classes", None)
         )
         self.fitness_function = SubsetSizePenalty(
             evaluator=evaluator,
