### 2025-04-27 d3e5adf
docs: update paper conclusion to cite regression task support
