+++ paper/paper.md
@@ -47,7 +47,7 @@ Despite the growing interest in cooperative co-evolutionary algorithms for featu
 
 `PyCCEA` addresses this gap by providing a well-organized, research-focused implementation of cooperative co-evolutionary algorithms tailored to feature selection. It incorporates widely used strategies from the literature and encourages standardization in experimental design and evaluation. By enabling consistent benchmarking and facilitating the development of new strategies, `PyCCEA` supports both methodological innovation and practical application in high-dimensional machine learning problems. Its release lowers the barrier to entry for researchers and practitioners, accelerating progress in the field of feature selection using evolutionary computation.
 
-Recent research has already leveraged `PyCCEA` as a foundational framework for implementing and evaluating cooperative co-evolutionary strategies in high-dimensional classification problems [@Venancio:2025], demonstrating its effectiveness in supporting reproducible and extensible scientific research. Planned enhancements aim to extend the framework’s generalization capabilities, enabling broader applicability to diverse problem domains, including regression and clustering tasks. These developments also aim to foster greater community engagement by making the framework more adaptable, accessible, and conducive to collaborative development.
+Recent research has already leveraged `PyCCEA` as a foundational framework for implementing and evaluating cooperative co-evolutionary strategies in high-dimensional classification problems [@Venancio:2025], demonstrating its effectiveness in supporting reproducible and extensible scientific research. To expand the framework’s usability beyond classification, we have recently introduced support for regression tasks, further broadening its applicability to diverse problem domains. These ongoing developments also seek to promote greater community engagement by making the framework increasingly adaptable, accessible, and conducive to collaborative research and development.
 
 # Acknowledgements
 
