### 2025-04-27 67cbfad
feat: add a classification dataset with 100k features
