### 2025-04-30 a67df1f
test: add unit test for Collaboration base class
