+++ tests/test_dataloader.py
@@ -1,7 +1,5 @@
-import os
 import pytest
 import logging
-import tempfile
 import numpy as np
 import pandas as pd
 from unittest.mock import patch
