### 2025-05-01 b41f5d7
test: add unit tests for random, elite and best collaboration
