### 2025-05-01 d984f5a
chore: add coverage log to stdout
