+++ tests/test_cooperation.py
@@ -25,7 +25,7 @@ def test_build_context_vector_concatenates_correctly() -> None:
 
 def test_get_best_collaborators(generation) -> None:
     """Test if the best collaborators are selected correctly."""
-    current_subpops, fitness, current_best, _ = generation
+    current_subpops, fitness, current_best, previous_subpops = generation
     # Test with a single subpopulation
     subpop_idx = 0
     indiv_idx = 0
@@ -37,11 +37,15 @@ def test_get_best_collaborators(generation) -> None:
         current_best=current_best
     )
 
-    expected_collaborators = [
-        np.array([1, 1, 0, 0, 1]),
-        np.array([1, 0]),
-        np.array([0, 0, 1])
-    ]
+    # Automatically compute expected collaborators
+    expected_collaborators = []
+    for i, subpop in enumerate(previous_subpops):
+        if i == subpop_idx:
+            expected_collaborators.append(current_subpops[i][indiv_idx])
+        else:
+            # Get index of individual with max fitness in subpop `i`
+            best_indiv_idx = np.argmax(fitness[i])
+            expected_collaborators.append(previous_subpops[i][best_indiv_idx])
 
     assert len(collaborators) == len(expected_collaborators)
     for i in range(len(collaborators)):
@@ -52,37 +56,36 @@ def test_get_elite_collaborators(generation) -> None:
     """Test if the elite collaborators are selected correctly."""
     current_subpops, fitness, _, previous_subpops = generation
 
-    # Define mocked return values for random.choices per subpopulation
-    # Assume we return the first individual from subpop[0] and the last one from subpop[2]
-    mock_returns = [
-        previous_subpops[0][0],
-        previous_subpops[2][-1],
-    ]
-
     # Get collaborators for the 1st (index = 0) individual of the 2nd (index = 1) subpopulation
     subpop_idx = 1
     indiv_idx = 0
-    sample_size = 3
+    sample_size = 2
 
     collaborator = SingleEliteCollaboration(sample_size=sample_size)
-    with patch("random.choices", side_effect=[[mock_returns[0]], [mock_returns[1]]]):
-        collaborators = collaborator.get_collaborators(
-            subpop_idx=subpop_idx,
-            indiv_idx=indiv_idx,
-            previous_subpops=previous_subpops,
-            current_subpops=current_subpops,
-            fitness=fitness
-        )
+    collaborators = collaborator.get_collaborators(
+        subpop_idx=subpop_idx,
+        indiv_idx=indiv_idx,
+        previous_subpops=previous_subpops,
+        current_subpops=current_subpops,
+        fitness=fitness
+    )
 
-    expected_collaborators = [
-        np.array([0, 1, 0, 1, 1]),  # mock_returns[0]
-        np.array([1, 1]),  # current_subpops[subpop_idx][indiv_idx]
-        np.array([0, 0, 1])  # mockl_returns[1]
-    ]
+    # Validate each collaborator is among the elite of the corresponding subpopulation
+    collaborator_idx = 0
+    for i in range(len(previous_subpops)):
+        if i == subpop_idx:
+            # Current subpop — ensure it is the same individual (indiv_idx)
+            assert np.array_equal(collaborators[collaborator_idx], current_subpops[subpop_idx][indiv_idx])
+        else:
+            # Other subpops — get indices of top-k (elite) based on fitness
+            subpop_fitness = fitness[i]
+            top_k_indices = np.argsort(subpop_fitness)[-sample_size:]
+            top_k_individuals = [previous_subpops[i][j] for j in top_k_indices]
 
-    assert len(collaborators) == len(expected_collaborators)
-    for i in range(len(collaborators)):
[…]
