### 2025-05-03 88ae2f6
test: improve unit tests for random collaboration strategy
