### 2025-05-03 08a1caf
test: ensure coverage for key CCEA components
