### 2025-05-03 8b90663
refactor: expose collaboration and decomposition strategies via __init__
