+++ pyccea/cooperation/__init__.py
@@ -1,0 +1,11 @@
+from .collaboration import Collaboration
+from .best import SingleBestCollaboration
+from .elite import SingleEliteCollaboration
+from .random import SingleRandomCollaboration
+
+__all__ = [
+    "Collaboration",
+    "SingleBestCollaboration",
+    "SingleEliteCollaboration",
+    "SingleRandomCollaboration",
+]
