+++ pyccea/decomposition/__init__.py
@@ -1,0 +1,15 @@
+from .grouping import FeatureGrouping
+from .dummy import DummyFeatureGrouping
+from .random import RandomFeatureGrouping
+from .ranking import RankingFeatureGrouping
+from .static import SequentialFeatureGrouping
+from .clustering import ClusteringFeatureGrouping
+
+__all__ = [
+    "FeatureGrouping",
+    "DummyFeatureGrouping",
+    "RandomFeatureGrouping",
+    "RankingFeatureGrouping",
+    "SequentialFeatureGrouping",
+    "ClusteringFeatureGrouping",
+]
