### 2025-05-03 e447ec7
test: add unit tests for the decomposition module
