+++ tests/test_decomposition.py
@@ -1,0 +1,147 @@
+import pytest
+import numpy as np
+from pyccea.decomposition import (
+    FeatureGrouping,
+    DummyFeatureGrouping,
+    RandomFeatureGrouping,
+    RankingFeatureGrouping,
+    SequentialFeatureGrouping,
+    ClusteringFeatureGrouping
+)
+
+
+def test_init_with_both_n_subcomps_and_subcomp_sizes() -> None:
+    """Test error is raised when both n_subcomps and subcomp_sizes are provided."""
+    with pytest.raises(AssertionError, match="Provide only one of the parameters"):
+        FeatureGrouping(n_subcomps=3, subcomp_sizes=[2, 3])
+
+
+def test_get_subcomponents_with_subcomp_sizes() -> None:
+    """Test the decomposition of features into subcomponents when subcomp_sizes is provided."""
+    X = np.array([[1, 2, 3, 4, 5, 6],
+                  [7, 8, 9, 10, 11, 12]])
+    feature_grouping = FeatureGrouping(subcomp_sizes=[2, 4])
+    subcomponents = feature_grouping._get_subcomponents(X)
+    assert len(subcomponents) == 2
+    assert np.array_equal(subcomponents[0], np.array([[1, 2], [7, 8]]))
+    assert np.array_equal(subcomponents[1], np.array([[3, 4, 5, 6], [9, 10, 11, 12]]))
+
+
+def test_get_subcomponents_with_n_subcomps() -> None:
+    """Test the decomposition of features into subcomponents when n_subcomps is provided."""
+    X = np.array([[1, 2, 3, 4, 5, 6],
+                  [7, 8, 9, 10, 11, 12]])
+    feature_grouping = FeatureGrouping(n_subcomps=3)
+    subcomponents = feature_grouping._get_subcomponents(X)
+    assert len(subcomponents) == 3
+    assert np.array_equal(subcomponents[0], np.array([[1, 2], [7, 8]]))
+    assert np.array_equal(subcomponents[1], np.array([[3, 4], [9, 10]]))
+    assert np.array_equal(subcomponents[2], np.array([[5, 6], [11, 12]]))
+
+
+def test_get_subcomponents_invalid_subcomp_sizes_shape() -> None:
+    """Test error is raised when subcomp_sizes do not match the number of features in X."""
+    X = np.random.rand(10, 5)  # 5 features
+    feature_grouping = FeatureGrouping(subcomp_sizes=[2, 2])  # Sum is 4, mismatch with 5
+
+    with pytest.raises(AssertionError, match="The sum of subcomponent sizes"):
+        feature_grouping._get_subcomponents(X)
+
+
+def test_dummy_decomposition() -> None:
+    """Test the decomposition of features using DummyFeatureGrouping."""
+    X = np.array([[1, 2, 3, 4, 5, 6],
+                  [7, 8, 9, 10, 11, 12]])
+    feature_grouping = DummyFeatureGrouping(subcomp_sizes=[4, 2], feature_idxs=np.array([5, 0, 2, 3, 4, 1]))
+    subcomponents, feature_idxs = feature_grouping.decompose(X)
+    assert len(subcomponents) == 2
+    assert np.array_equal(subcomponents[0], np.array([[6, 1, 3, 4], [12, 7, 9, 10]]))
+    assert np.array_equal(subcomponents[1], np.array([[5, 2], [11, 8]]))
+    assert np.array_equal(feature_idxs, np.array([5, 0, 2, 3, 4, 1]))
+
+
+def test_random_decomposition() -> None:
+    """Test the decomposition of features using RandomFeatureGrouping."""
[…]
