### 2025-05-03 399bda7
feat: add raises for small sample sizes and invalid alternative in statistical comparison
