### 2025-05-03 865985b
test: add unit tests for statistical comparison function
