### 2025-05-20 9b4971c
feat: add unit tests for fitness functions
