+++ pyccea/fitness/__init__.py
@@ -1,0 +1,9 @@
+from .distance import DistanceBasedFitness
+from .penalty import SubsetSizePenalty
+from .function import WrapperFitnessFunction
+
+__all__ = [
+    "DistanceBasedFitness",
+    "SubsetSizePenalty",
+    "WrapperFitnessFunction"
+]
