+++ tests/test_fitness.py
@@ -1,0 +1,214 @@
+import pytest
+import numpy as np
+from unittest.mock import MagicMock
+from pyccea.fitness import (
+    DistanceBasedFitness,
+    SubsetSizePenalty,
+    WrapperFitnessFunction
+)
+
+
+def test_evaluate_predictive_performance(dummy_dataloader, dummy_classification_evaluator) -> None:
+    """Test the evaluate predictive performance method."""
+    fitness = WrapperFitnessFunction(evaluator=dummy_classification_evaluator)
+    context_vector = np.array([1, 0, 1, 0, 1])
+    _ = fitness._evaluate_predictive_performance(
+        context_vector=context_vector,
+        data=dummy_dataloader
+    )
+    dummy_classification_evaluator.evaluate.assert_called_once()
+
+
+def test_subset_size_penalty_weights_sum(dummy_classification_evaluator) -> None:
+    """Test that an AssertionError is raised if the sum of weights is not 1 for penalty-based fitness."""
+    with pytest.raises(AssertionError, match="must be 1"):
+        SubsetSizePenalty(dummy_classification_evaluator, [0.7, 0.4])
+
+
+def test_subset_size_penalty_weights_length(dummy_classification_evaluator) -> None:
+    """Test that an AssertionError is raised if the number of weights is not 2 for penalty-based fitness."""
+    with pytest.raises(AssertionError, match="only two components"):
+        SubsetSizePenalty(dummy_classification_evaluator, [0.7, 0.3, 0.2])
+
+
+def test_subset_size_penalty_classification_evaluation(
+    dummy_classification_evaluator,
+    dummy_dataloader
+) -> None:
+    """Test the evaluate method of penalty-based fitness in classification."""
+    weights = [0.7, 0.3]
+    fitness = SubsetSizePenalty(dummy_classification_evaluator, weights)
+    
+    # Mock the _evaluate_predictive_performance method
+    fitness._evaluate_predictive_performance = MagicMock(
+        return_value={dummy_classification_evaluator.eval_function: 0.8}
+    )
+
+    context_vector = np.array([1, 1, 1, 1, 0, 0, 0, 0, 0, 0])
+    fitness_value = fitness.evaluate(context_vector, dummy_dataloader)
+
+    expected_penalty = context_vector.sum() / dummy_dataloader.n_features
+    expected_evaluation = 0.8
+    expected_fitness = 1 * weights[0] * expected_evaluation - weights[1] * expected_penalty
+
+    assert pytest.approx(fitness_value) == expected_fitness
+    fitness._evaluate_predictive_performance.assert_called_once_with(context_vector, dummy_dataloader)
+
+
+def test_subset_size_penalty_regression_evaluation(
+    dummy_regression_evaluator,
+    dummy_dataloader
+) -> None:
+    """Test the evaluate method of penalty-based fitness in regression."""
+    weights = [0.7, 0.3]
+    fitness = SubsetSizePenalty(dummy_regression_evaluator, weights)
+    
+    # Mock the _evaluate_predictive_performance method
+    fitness._evaluate_predictive_performance = MagicMock(
+        return_value={dummy_regression_evaluator.eval_function: 0.8}
+    )
+
+    context_vector = np.array([1, 1, 1, 1, 0, 0, 0, 0, 0, 0])
+    fitness_value = fitness.evaluate(context_vector, dummy_dataloader)
+
+    expected_penalty = context_vector.sum() / dummy_dataloader.n_features
+    expected_evaluation = 0.8
+    expected_fitness = -1 * weights[0] * expected_evaluation - weights[1] * expected_penalty
+
[…]
