+++ tests/test_stats.py
@@ -5,8 +5,9 @@ from pyccea.utils.stats import statistical_comparison_between_independent_sample
 
 def test_ttest_used_for_normal_data() -> None:
     """Test if t-test is used for normally distributed data."""
-    x = np.random.normal(loc=80, scale=10, size=30)
-    y = np.random.normal(loc=90, scale=5, size=30)
+    np.random.seed(42)
+    x = np.random.normal(loc=80, scale=5, size=100)
+    y = np.random.normal(loc=90, scale=5, size=100)
     test_name, _, p_comparison, reject_null = statistical_comparison_between_independent_samples(x, y, alpha=0.05)
     assert test_name == "t-test"
     assert reject_null is True
@@ -15,6 +16,7 @@ def test_ttest_used_for_normal_data() -> None:
 
 def test_mann_whitney_u_used_for_non_normal_data() -> None:
     """Test if Mann-Whitney U test is used for non-normally distributed data."""
+    np.random.seed(42)
     x = np.random.exponential(scale=1.00, size=30)
     y = np.random.exponential(scale=1.05, size=30)
     test_name, _, p_comparison, reject_null = statistical_comparison_between_independent_samples(x, y, alpha=0.05)
@@ -25,6 +27,7 @@ def test_mann_whitney_u_used_for_non_normal_data() -> None:
 
 def test_small_sample_size() -> None:
     """Test that a ValueError is raised for small sample sizes (lower than 3)."""
+    np.random.seed(42)
     x = np.random.normal(loc=80, scale=10, size=2)
     y = np.random.normal(loc=90, scale=5, size=2)
     with pytest.raises(ValueError, match="1st sample has only 2 observation\\(s\\)\\.|2nd sample has only 2 observation\\(s\\)\\."):
@@ -36,6 +39,7 @@ def test_small_sample_size() -> None:
 
 def test_invalid_alternative() -> None:
     """Test that a ValueError is raised for invalid alternative hypothesis."""
+    np.random.seed(42)
     x = np.random.normal(loc=80, scale=10, size=30)
     y = np.random.normal(loc=90, scale=5, size=30)
     with pytest.raises(ValueError, match="Alternative hypothesis must be 'two-sided', 'greater', or 'less'."):
