### 2025-05-20 cec9a97
test: add preliminary tests for initialization methods
