+++ pyccea/initialization/__init__.py
@@ -1,2 +1,9 @@
+from .build import SubpopulationInitialization
 from .binary import RandomBinaryInitialization
 from .continuous import RandomContinuousInitialization
+
+__all__ = [
+    "SubpopulationInitialization",
+    "RandomBinaryInitialization",
+    "RandomContinuousInitialization"
+]
