### 2025-05-21 8024c4c
test: improve unit tests for initialization methods
