### 2025-05-21 df8c82c
test: finish unit tests for initialization methods
