### 2025-05-21 85e9d78
test: add unit tests for the binary genetic algorithm
