+++ tests/test_optimizers.py
@@ -1,0 +1,307 @@
+import pytest
+import numpy as np
+from pyccea.optimizers import (
+    BinaryGeneticAlgorithm,
+    DifferentialEvolution
+)
+
+
+def test_binary_genetic_algorithm_initialization(optimizer_config, dummy_dataloader) -> None:
+    """Test the initialization of the BinaryGeneticAlgorithm class."""
+    # Create an instance of BinaryGeneticAlgorithm
+    optimizer = BinaryGeneticAlgorithm(
+        subpop_size=optimizer_config["coevolution"]["subpop_sizes"][0],
+        n_features=dummy_dataloader.n_features,
+        conf=optimizer_config
+    )
+    # Check that the attributes are set correctly
+    assert optimizer.n_features == dummy_dataloader.n_features
+    assert optimizer.subpop_size == optimizer_config["coevolution"]["subpop_sizes"][0]
+    assert optimizer.mutation_rate == optimizer_config["optimizer"]["mutation_rate"]
+    assert optimizer.crossover_rate == optimizer_config["optimizer"]["crossover_rate"]
+    assert optimizer.selection_method == optimizer_config["optimizer"]["selection_method"]
+
+
+def test_binary_genetic_algorithm_invalid_initialization(optimizer_config, dummy_dataloader) -> None:
+    """Test the invalid initialization of the BinaryGeneticAlgorithm class."""
+    # Test with invalid selection method
+    with pytest.raises(AssertionError):
+        BinaryGeneticAlgorithm(
+            subpop_size=optimizer_config["coevolution"]["subpop_sizes"][0],
+            n_features=dummy_dataloader.n_features,
+            conf={"optimizer": {**optimizer_config["optimizer"], "selection_method": "invalid_method"}}
+        )
+    # Test with invalid mutation rate
+    with pytest.raises(ValueError):
+        BinaryGeneticAlgorithm(
+            subpop_size=optimizer_config["coevolution"]["subpop_sizes"][0],
+            n_features=dummy_dataloader.n_features,
+            conf={"optimizer": {**optimizer_config["optimizer"], "mutation_rate": -0.1}}
+        )
+    # Test with invalid crossover rate
+    with pytest.raises(ValueError):
+        BinaryGeneticAlgorithm(
+            subpop_size=optimizer_config["coevolution"]["subpop_sizes"][0],
+            n_features=dummy_dataloader.n_features,
+            conf={"optimizer": {**optimizer_config["optimizer"], "crossover_rate": 1.5}}
+        )
+
+
+def test_single_point_crossover(monkeypatch, optimizer_config) -> None:
+    """Test _single_point_crossover with mocked crossover point and probability."""
+    optimizer = BinaryGeneticAlgorithm(
+        subpop_size=10,
+        n_features=5,
+        conf=optimizer_config
+    )
+    crossover_point = 3
+    optimizer_config["optimizer"]["crossover_rate"] = 0.10
+    parent_a = np.array([1, 0, 1, 0, 1])
+    parent_b = np.array([0, 1, 0, 1, 0])
+
+    # Mock np.random.uniform to always return 0.1 (less than crossover_rate)
+    monkeypatch.setattr("numpy.random.uniform", lambda *a, **k: optimizer_config["optimizer"]["crossover_rate"])
+    # Mock np.random.randint to always return 3 as crossover point
+    monkeypatch.setattr("numpy.random.randint", lambda *a, **k: crossover_point)
+
[…]
