+++ pyccea/optimizers/__init__.py
@@ -1,0 +1,7 @@
+from .genetic_algorithm import BinaryGeneticAlgorithm
+from .differential_evolution import DifferentialEvolution
+
+__all__ = [
+    "BinaryGeneticAlgorithm",
+    "DifferentialEvolution"
+]
