### 2025-05-21 007782f
feat: add unit tests for the differential evolution
