+++ tests/test_optimizers.py
@@ -305,3 +305,214 @@ def test_binary_genetic_algorithm_steady_state_evolve(monkeypatch, optimizer_con
     assert uniform_calls["count"] == 4  # 2 per offspring (crossover + mutation), 2 offspring
 
 
+def test_differential_evolution_initialization(optimizer_config, dummy_dataloader) -> None:
+    """Test the initialization of the DifferentialEvolution class."""
+    optimizer_config["optimizer"]["method"] = "DE"
+    optimizer = DifferentialEvolution(
+        subpop_size=optimizer_config["coevolution"]["subpop_sizes"][0],
+        n_features=dummy_dataloader.n_features,
+        conf=optimizer_config
+    )
+    # Check that the attributes are set correctly
+    assert optimizer.n_features == dummy_dataloader.n_features
+    assert optimizer.subpop_size == optimizer_config["coevolution"]["subpop_sizes"][0]
+    assert optimizer.scaling_factor == optimizer_config["optimizer"]["scaling_factor"]
+    assert optimizer.crossover_probability == optimizer_config["optimizer"]["crossover_probability"]
+
+
+def test_differential_evolution_invalid_initialization(optimizer_config, dummy_dataloader) -> None:
+    """Test the invalid initialization of the DifferentialEvolution class."""
+    optimizer_config["optimizer"]["method"] = "DE"
+    # Test with invalid scaling factor
+    with pytest.raises(ValueError):
+        DifferentialEvolution(
+            subpop_size=optimizer_config["coevolution"]["subpop_sizes"][0],
+            n_features=dummy_dataloader.n_features,
+            conf={"optimizer": {**optimizer_config["optimizer"], "scaling_factor": 3}}
+        )
+    # Test with invalid crossover probability
+    with pytest.raises(ValueError):
+        DifferentialEvolution(
+            subpop_size=optimizer_config["coevolution"]["subpop_sizes"][0],
+            n_features=dummy_dataloader.n_features,
+            conf={"optimizer": {**optimizer_config["optimizer"], "crossover_probability": -0.1}}
+        )
+    # Test with invalid subpopulation size
+    with pytest.raises(ValueError):
+        DifferentialEvolution(
+            subpop_size=3,
+            n_features=dummy_dataloader.n_features,
+            conf=optimizer_config
+        )
+
+
+def test_select_solutions(optimizer_config) -> None:
+    """Test the _select_solutions method of the DifferentialEvolution class."""
+    optimizer_config["optimizer"]["method"] = "DE"
+    optimizer = DifferentialEvolution(
+        subpop_size=5,
+        n_features=3,
+        conf=optimizer_config
+    )
+    # Create a population with 5 individuals
+    population = np.array([
+        [1.5, 0.2, 1.0],
+        [0.1, 1.0, 0.5],
+        [1.0, 0.5, 0.2],
+        [0.3, -0.8, 1.2],
+        [-1.2, 0.4, 0.9]
+    ])
+    target_vector_idx = 2
+    selected_solutions = optimizer._select_solutions(population, target_vector_idx)
+    # Check that the selected solutions are different from the target vector
+    for solution in selected_solutions:
+        assert not np.array_equal(solution, population[target_vector_idx])
+    # Check that the selected solutions are different from each other
+    for i in range(len(selected_solutions)):
+        for j in range(i + 1, len(selected_solutions)):
[…]
