### 2025-05-26 a7b2766
feat: support predefined data folds for model selection
