+++ tests/test_dataloader.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 from unittest.mock import patch
 from pyccea.utils.datasets import DataLoader
-from sklearn.model_selection import KFold, LeaveOneOut, StratifiedKFold
+from sklearn.model_selection import GroupKFold, KFold, LeaveOneOut, StratifiedKFold
 
 
 # Disable logging output during tests
@@ -331,6 +331,46 @@ def test_standard_normalization(complete_data_conf: dict) -> None:
     assert np.allclose(loader.X_train.std(axis=0), np.ones(loader.X_train.shape[1]))
 
 
+def test_kfold_model_selection_prefold_missing_fold_column(monkeypatch, complete_data_conf: dict) -> None:
+    """Test k-fold model selection when 'prefold' is True but the dataset lacks the required 'fold' column."""
+    complete_data_conf["splitter"]["prefold"] = True
+    complete_data_conf["splitter"]["kfolds"] = 3
+    loader = DataLoader("dummy", complete_data_conf)
+    monkeypatch.setitem(DataLoader.DATASETS, "dummy", {"task": "classification"})
+    loader.data = pd.DataFrame({
+        "0": [1, 2, 3, 4, 5],
+        "1": [6, 7, 8, 9, 0],
+        "label": [0, 1, 0, 1, 1],
+        "subset": ["train", "train", "train", "test", "test"]
+    })
+    loader.X = loader._get_input()
+    loader.y = loader._get_output()
+    loader._split()
+    with pytest.raises(AssertionError, match="The 'fold' column should be specified"):
+        loader._model_selection()
+
+
+def test_kfold_model_selection_prefold_mismatched_folds(monkeypatch, complete_data_conf: dict) -> None:
+    """Test k-fold model selection with 'prefold' set to True and kfolds specified does not match the number of
+    folds in the dataset."""
+    complete_data_conf["splitter"]["prefold"] = True
+    complete_data_conf["splitter"]["kfolds"] = 4
+    loader = DataLoader("dummy", complete_data_conf)
+    monkeypatch.setitem(DataLoader.DATASETS, "dummy", {"task": "classification"})
+    loader.data = pd.DataFrame({
+        "0": [1, 2, 3, 4, 5],
+        "1": [6, 7, 8, 9, 0],
+        "label": [0, 1, 0, 1, 1],
+        "subset": ["train", "train", "train", "test", "test"],
+        "fold": [1, 2, 3, 1, 2]
+    })
+    loader.X = loader._get_input()
+    loader.y = loader._get_output()
+    loader._split()
+    with pytest.raises(AssertionError, match="The number of folds in the training set"):
+        loader._model_selection()
+
+
 def test_stratified_kfold_model_selection(monkeypatch, complete_data_conf: dict) -> None:
     """Test _model_selection with k_fold splitter."""
     complete_data_conf["splitter"]["kfolds"] = 2
@@ -361,10 +401,94 @@ def test_kfold_model_selection(monkeypatch, complete_data_conf: dict) -> None:
     assert isinstance(loader.splitter, KFold)
 
 
+def test_group_kfold_model_selection(monkeypatch, complete_data_conf: dict) -> None:
+    """Test k-fold model selection when 'prefold' is True."""
+    complete_data_conf["splitter"]["prefold"] = True
+    del complete_data_conf["splitter"]["stratified"]
+    del complete_data_conf["splitter"]["kfolds"]
+    loader = DataLoader("dummy", complete_data_conf)
[…]
