+++ pyccea/utils/datasets.py
@@ -1,12 +1,13 @@
 import os
 import toml
 import logging
+import warnings
 import numpy as np
 import pandas as pd
 import importlib.resources
 from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import MinMaxScaler, StandardScaler
-from sklearn.model_selection import KFold, LeaveOneOut, StratifiedKFold
+from sklearn.model_selection import GroupKFold, KFold, LeaveOneOut, StratifiedKFold
 
 
 class DataLoader():
@@ -107,29 +108,35 @@ class DataLoader():
     def _parse_splitter_parameters(self) -> None:
         """Parse parameters from the splitter section of the data configuration file."""
         if self.splitter_type == "k_fold":
-            if "kfolds" not in self.conf["splitter"]:
+            if ("kfolds" not in self.conf["splitter"]) and (self.conf["splitter"].get("prefold", False) is False):
                 raise AssertionError(
                     "The parameter 'kfolds' should be specified in the splitter section of the "
-                    "data configuration file when 'splitter_type' is set to 'k_fold'."
+                    "data configuration file when 'splitter_type' is set to 'k_fold' and 'prefold' "
+                    "is set to False or is not defined in the splitter section."
                 )
-            self.kfolds = self.conf["splitter"]["kfolds"]
+            self.kfolds = self.conf["splitter"].get("kfolds")
             self.stratified = self.conf["splitter"].get("stratified", False)
         if self.splitter_type == "leave_one_out":
-            logging.warning(
-                "Careful! It is not recommended to use Leave-One-Out when your problem has large "
-                "datasets or costly machine learning models to fit."
-            )
             if "kfolds" in self.conf["splitter"]:
-                logging.warning(
+                warnings.warn(
                     "You specified the number of folds using Leave-One-Out (LOO). However, LOO is"
                     " equivalent to K-Fold when K is equal to the number of examples. Therefore, "
-                    "the value of 'kfolds' parameter will be ignored in this case."
+                    "the value of 'kfolds' parameter will be ignored in this case.",
+                    UserWarning
                 )
             if "stratified" in self.conf["splitter"]:
-                logging.warning(
+                warnings.warn(
                     "You specified the 'stratified' parameter using Leave-One-Out (LOO). However," 
                     " the validation folds made by the LOO have only one sample. Therefore, the "
-                    "value of 'stratified' parameter will be ignored in this case."
+                    "value of 'stratified' parameter will be ignored in this case.",
+                    UserWarning
+                )
+            if self.conf["splitter"].get("prefold"):
+                warnings.warn(
+                    "You specified the 'prefold' parameter using Leave-One-Out (LOO). However, "
+                    "the validation folds made by the LOO have only one sample. Therefore, the "
+                    "value of 'prefold' parameter will be ignored in this case.",
+                    UserWarning
                 )
         self.preset = self.conf["splitter"].get("preset", False)
         if self.preset:
@@ -152,6 +159,7 @@ class DataLoader():
                     "The 'test_ratio' parameter should be within the range of 0 and 1, excluding "
                     "extreme values (i.e., 0 < 'test_ratio' < 1)."
                 )
+        self.prefold = self.conf["splitter"].get("prefold", False)
 
     def _parse_normalization_parameters(self) -> None:
         """Parse parameters from the normalization section of the data configuration file."""
@@ -215,9 +223,10 @@ class DataLoader():
         X : pd.DataFrame (n_examples, n_features)
             Input data (features).
         """
-        selected_cols = set(self.data.columns) - set(["label", "subset", "fold"])
-        selected_cols = sorted([int(col) for col in selected_cols])
-        X = self.data.iloc[:, selected_cols].copy()
[…]
