### 2025-05-26 09486ea
chore: update package versions of scikit-learn and imbalanced-learn
