### 2025-06-03 3e56da1
chore: update PyCCEA version from 0.1.0 to 0.2.0
