+++ setup.py
@@ -1,16 +1,23 @@
-from setuptools import setup, find_packages
+import os
 
+from setuptools import find_packages, setup
 
 with open("README.md") as f:
     long_description = f.read()
 
-with open("requirements.txt") as f:
-    requirements = [line.strip() for line in f.readlines()]
+requirements_path = os.path.join(os.path.dirname(__file__), "requirements.txt")
+if os.path.exists(requirements_path):
+    with open(requirements_path, encoding="utf-8") as f:
+        requirements = [
+            line.strip() for line in f if line.strip() and not line.startswith("#")
+        ]
+else:
+    requirements = []
 
 
 setup(
     name="pyccea",
-    version="0.1.0",
+    version="0.2.0",
     description="Cooperative co-evolutionary algorithms for feature selection in high-dimensional data",
     long_description=long_description,
     long_description_content_type="text/markdown",
@@ -19,7 +26,7 @@ setup(
     url="https://github.com/pedbrgs/PyCCEA",
     packages=find_packages(include=["pyccea*"]),
     package_data={
-        "pyccea": ["datasets/*.csv", "parameters/*.toml"],
+        "pyccea": ["datasets/*.parquet", "parameters/*.toml"],
     },
     install_requires=requirements,
     classifiers=[
