### 2025-06-03 5109599
fix: add fsspec to dependencies
