+++ requirements.txt
@@ -13,3 +13,4 @@ pytest-cov==6.1.1
 pytest-asyncio==0.26.0
 pytest-deadfixtures==2.2.1
 pytest-mock==3.14.0
+fsspec==2025.5.1
