### 2025-06-03 c0ea887
chore: update PyCCEA version to 1.0.0
