### 2025-06-03 248d032
chore: reduce number of generations by default
