### 2025-06-03 220b823
feat: make upper bounds for self-tuning configurable in CCPSTFG
