### 2025-06-03 623ce1d
fix: Agglomerative Clustering parameters after package upgrade
