### 2025-06-04 b6f74db
chore: change some RF and CCPSTFG parameters
