+++ pyccea/parameters/ccpstfg.toml
@@ -1,7 +1,7 @@
 # This file specifies the parameters of the Cooperative Co-Evolutionary Algorithm with Projection-based Self-Tuning Feature Grouping (CCPSTFG)
 
 [coevolution]
-subpop_sizes = [100]
+subpop_sizes = [30]
 max_gen = 10000
 max_gen_without_improvement = 100
 optimized_resource_allocation = true
@@ -12,7 +12,7 @@ drop = true
 max_n_clusters = 10
 max_n_pls_components = 30
 removal_quantile_step_size = 0.05
-max_removal_quantile = 1.00
+max_removal_quantile = 0.95
 clustering_model_type = "agglomerative_clustering"
 
 [collaboration]
