### 2025-06-04 #8348 @editorialbot
[REVIEW]: PyCCEA: A Python package of cooperative co-evolutionary algorithms for feature selection in high-dimensional data
**Submitting author:** @pedbrgs (<a href="http://orcid.org/0000-0003-4665-562X">Pedro Vinícius A. B. de Venâncio</a>)
**Repository:** https://github.com/pedbrgs/PyCCEA/
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.2
