### 2025-06-04 #8348 @bkrayfield
Thank you, @geoffreyp and @maximtrp, for agreeing to review this JOSS submission.

If this is your first JOSS review, the process begins with a checklist. You can generate it by commenting: "(at) editorialbot generate my checklist"

Use this checklist as your guide, opening issues in the repository for any comments or questions and linking them back to the review thread.

JOSS recommends creating the initial set of issues within 4–6 weeks. I’ll check in periodically to ensure everything is on track.

Reviewer guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Please let me know if you have any questions, I’m happy to help.
