### 2025-06-05 45f72e6
doc: add CCPSTFG to the list of available CCEAs
