+++ pyccea/coevolution/README.md
@@ -4,6 +4,7 @@ The following cooperative co-evolutionary algorithms (CCEAs) are currently imple
 
 | CCEA | Reference |
 |------|--------|
-|  CCEAFS   |   [Rashid, A. N. M., et al. "A novel penalty-based wrapper objective function for feature selection in Big Data using cooperative co-evolution." IEEE Access 8 (2020): 150113-150129.](https://ieeexplore.ieee.org/abstract/document/9167222/)     |
-|  CCFSRFG1    |  [Rashid, A. N. M., et al. "Cooperative co-evolution for feature selection in Big Data with random feature grouping." Journal of Big Data 7.1 (2020): 1-42.](https://journalofbigdata.springeropen.com/articles/10.1186/s40537-020-00381-y)      |
-|  CCFSRFG2    |  [Rashid, A. N. M., et al. "Cooperative co-evolution for feature selection in Big Data with random feature grouping." Journal of Big Data 7.1 (2020): 1-42.](https://journalofbigdata.springeropen.com/articles/10.1186/s40537-020-00381-y)      |
+|  CCEAFS   |   [Rashid, A. N. M., et al. "A novel penalty-based wrapper objective function for feature selection in Big Data using cooperative co-evolution." IEEE Access (2020).](https://ieeexplore.ieee.org/abstract/document/9167222/)     |
+|  CCFSRFG1    |  [Rashid, A. N. M., et al. "Cooperative co-evolution for feature selection in Big Data with random feature grouping." Journal of Big Data (2020).](https://journalofbigdata.springeropen.com/articles/10.1186/s40537-020-00381-y)      |
+|  CCFSRFG2    |  [Rashid, A. N. M., et al. "Cooperative co-evolution for feature selection in Big Data with random feature grouping." Journal of Big Data (2020).](https://journalofbigdata.springeropen.com/articles/10.1186/s40537-020-00381-y)      |
+|  CCPSTFG    |  [Venâncio, P. V. A. B., et al. "A self-tuning decomposition strategy in cooperative co-evolutionary algorithms for high-dimensional feature selection." Knowledge-Based Systems (2025).](https://journalofbigdata.springeropen.com/articles/10.1186/s40537-020-00381-y)      |
