### 2025-06-06 #8348 @maximtrp
@pedbrgs Please correct me if I am wrong, but there is no documentation for your package now. If so, please create it (e.g. on [readthedocs.org](https://readthedocs.org))
