### 2025-06-06 #8348 @maximtrp
I also strongly suggest using a GitHub workflow for publishing your package on PyPi. It's also a good practice to post release information on GitHub with a change log.
