### 2025-06-06 #8348 @geoffreyp
Hello,

To start, here are some issues I encountered:
- Which version of Python should we use?
- I was unable to test the example from your README by following your instructions (pip install pyccea):
- How can we contribute to the project? The CONTRIB file is missing ([The github documentation](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)).

```sh
(env) (base) ➜ review_joss python test.py
Traceback (most recent call last):
  File "/Users/geoffreypruvost/Projets/review_joss/test.py", line 1, in <module>
    import toml
ModuleNotFoundError: No module named 'toml'
```

In general, the README is not sufficient as documentation. We need:
- a very detailed user documentation, guiding the user step-by-step from installation to usage (using GitHub's wiki system or Read the
- a documentation of classes and methods for developers, or for users who want more customization for their use.
