### 2025-06-06 fedf9e1
feat: update packaging configuration to use pyproject.toml as recommended
