### 2025-06-07 84ad17f
docs: add contributing guidelines
