### 2025-06-07 aff69c8
docs: add python version requirement
