### 2025-06-08 9fdd60b
docs: set up sphinx docs and github actions for auto-deploy
