### 2025-06-08 1d71d0b
docs: add publish branch in the documentation worklflow
