+++ .github/workflows/docs.yml
@@ -25,8 +25,9 @@ jobs:
           cd docs
           make html
 
-      - name: Deploy to GitHub Pages
+        - name: Deploy to GitHub Pages
         uses: peaceiris/actions-gh-pages@v3
         with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: ./docs/build/html
+            github_token: ${{ secrets.GITHUB_TOKEN }}
+            publish_dir: ./docs/build/html
+            publish_branch: gh-pages
