### 2025-06-08 c7f9797
fix: yaml syntax
