+++ .github/workflows/docs.yml
@@ -25,7 +25,7 @@ jobs:
           cd docs
           make html
 
-        - name: Deploy to GitHub Pages
+      - name: Deploy to GitHub Pages
         uses: peaceiris/actions-gh-pages@v3
         with:
             github_token: ${{ secrets.GITHUB_TOKEN }}
