### 2025-06-08 06f0f7d
fix: correct publish_dir path and working directory in the docs workflow
