+++ _sources/index.rst.txt
@@ -1,17 +1,0 @@
-.. PyCCEA documentation master file, created by
-   sphinx-quickstart on Sun Jun  8 09:12:50 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-PyCCEA documentation
-====================
-
-Add your content using ``reStructuredText`` syntax. See the
-`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
-documentation for details.
-
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
