### 2025-06-08 a0bbad4
fix: build and deploy docs
