### 2025-06-08 c3f6c0c
docs: add first doc sections
