+++ docs/source/conf.py
@@ -1,31 +1,42 @@
-# Configuration file for the Sphinx documentation builder.
-#
-# For the full list of built-in configuration values, see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
+import os
+import sys
+sys.path.insert(0, os.path.abspath("../../"))
 
-# -- Project information -----------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
-
-project = 'PyCCEA'
-copyright = '2025, Pedro Vinícius Almeida Borges de Venâncio'
-author = 'Pedro Vinícius Almeida Borges de Venâncio'
-release = '1.0.1'
-
-# -- General configuration ---------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
+project = "PyCCEA"
+copyright = "2025, pedbrgs"
+author = "Pedro Vinícius Almeida Borges de Venâncio"
 
 extensions = [
-    'sphinx.ext.autodoc',
-    'numpydoc'
+    "sphinx.ext.autodoc",
+    "numpydoc",
+    "sphinx.ext.viewcode",
 ]
 
-templates_path = ['_templates']
+templates_path = ["_templates"]
 exclude_patterns = []
 
+html_theme = "sphinx_rtd_theme"
 
-
-# -- Options for HTML output -------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
+numpydoc_show_class_members = True
 
 html_theme = 'sphinx_rtd_theme'
+
+html_logo = '../figures/logo.png'
+
 html_static_path = ['_static']
+
+html_css_files = [
+    'custom.css',
+]
+
+html_favicon = '../figures/favicon.png'
+
+html_theme_options = {
+    'logo_only': True,
+    'style_nav_header_background': '#2980B9',
+    'collapse_navigation': True,
+    'sticky_navigation': False,
+    'navigation_depth': 4,
+    'includehidden': True,
+    'titles_only': False
+}
