+++ docs/source/pages/contributing.rst
@@ -1,0 +1,14 @@
+Contributing
+============
+
+We welcome contributions to PyCCEA!
+
+There are several ways you can get involved:
+
+- **Report issues** or suggest features by opening `issues <https://github.com/pedbrgs/pyccea/issues>`_ on the GitHub repository.
+- **Submit pull requests** to help improve the code, fix bugs, or add new features.
+- **Contact Pedro Vinícius** directly if you have questions or want to discuss contributions.
+
+When you submit a pull request or reach out via email, `Pedro Vinícius <https://github.com/pedbrgs>`_ will review and respond as soon as possible.
+
+Thank you for helping make PyCCEA better!
