+++ _sources/index.rst.txt
@@ -1,17 +1,13 @@
-.. PyCCEA documentation master file, created by
-   sphinx-quickstart on Sun Jun  8 09:12:50 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-PyCCEA documentation
+Documentation
 ====================
 
-Add your content using ``reStructuredText`` syntax. See the
-`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
-documentation for details.
-
+PyCCEA is an open-source Python package for feature selection in high-dimensional datasets using cooperative co-evolutionary algorithms (CCEAs). It provides a modular framework that enables researchers and practitioners to implement, evaluate, and extend CCEAs through interchangeable components such as decomposition strategies, fitness evaluators, and optimizers. Designed with reproducibility and experimentation in mind, PyCCEA supports wrapper-based evaluations with scikit-learn models and includes benchmark datasets and baseline algorithms for comparative analysis in classification and regression tasks.
 
 .. toctree::
-   :maxdepth: 2
-   :caption: Contents:
+   :maxdepth: 1
+   :numbered: 4
+   :caption: Contents
 
+   pages/installation
+   pages/quickstart
+   pages/contributing
