+++ _sources/pages/installation.rst.txt
@@ -1,0 +1,16 @@
+Installation
+============
+
+To install the PyCCEA package directly from PyPI, use the following command in a Python ≥ 3.10 environment:
+
+.. code-block:: bash
+
+    pip install pyccea
+
+Alternatively, to install the latest development version directly from GitHub, run:
+
+.. code-block:: bash
+
+    pip install git+https://github.com/pedbrgs/pyccea.git
+
+Make sure you have `pip` installed and an active internet connection to download dependencies.
