### 2025-06-08 43df820
docs: add API Reference and Citation sections
