+++ docs/source/conf.py
@@ -10,7 +10,11 @@ extensions = [
     "sphinx.ext.autodoc",
     "numpydoc",
     "sphinx.ext.viewcode",
+    'sphinx.ext.autosummary',
 ]
+
+autosummary_generate = True
+
 
 templates_path = ["_templates"]
 exclude_patterns = []
