+++ docs/source/pages/modules.rst
@@ -1,0 +1,18 @@
+API Reference
+=============
+
+The following is the API reference for the PyCCEA package.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Modules
+
+   pyccea.coevolution
+   pyccea.cooperation
+   pyccea.decomposition
+   pyccea.evaluation
+   pyccea.fitness
+   pyccea.initialization
+   pyccea.optimizers
+   pyccea.projection
+   pyccea.utils
