### 2025-06-08 367920f
fix: generate module-level .rst files
