+++ _sources/pages/pyccea.decomposition.ranking.rst.txt
@@ -1,0 +1,7 @@
+pyccea.decomposition.ranking module
+===================================
+
+.. automodule:: pyccea.decomposition.ranking
+   :members:
+   :show-inheritance:
+   :undoc-members:
