+++ _sources/pages/pyccea.projection.cipls.rst.txt
@@ -1,0 +1,7 @@
+pyccea.projection.cipls module
+==============================
+
+.. automodule:: pyccea.projection.cipls
+   :members:
+   :show-inheritance:
+   :undoc-members:
