+++ _sources/pages/pyccea.decomposition.random.rst.txt
@@ -1,0 +1,7 @@
+pyccea.decomposition.random module
+==================================
+
+.. automodule:: pyccea.decomposition.random
+   :members:
+   :show-inheritance:
+   :undoc-members:
