+++ _sources/pages/pyccea.decomposition.static.rst.txt
@@ -1,0 +1,7 @@
+pyccea.decomposition.static module
+==================================
+
+.. automodule:: pyccea.decomposition.static
+   :members:
+   :show-inheritance:
+   :undoc-members:
