+++ _sources/pages/pyccea.coevolution.ccfsrfg1.rst.txt
@@ -1,0 +1,7 @@
+pyccea.coevolution.ccfsrfg1 module
+==================================
+
+.. automodule:: pyccea.coevolution.ccfsrfg1
+   :members:
+   :show-inheritance:
+   :undoc-members:
