+++ _sources/pages/pyccea.evaluation.wrapper.rst.txt
@@ -1,0 +1,7 @@
+pyccea.evaluation.wrapper module
+================================
+
+.. automodule:: pyccea.evaluation.wrapper
+   :members:
+   :show-inheritance:
+   :undoc-members:
