+++ _sources/pages/pyccea.decomposition.grouping.rst.txt
@@ -1,0 +1,7 @@
+pyccea.decomposition.grouping module
+====================================
+
+.. automodule:: pyccea.decomposition.grouping
+   :members:
+   :show-inheritance:
+   :undoc-members:
