+++ _sources/pages/pyccea.initialization.continuous.rst.txt
@@ -1,0 +1,7 @@
+pyccea.initialization.continuous module
+=======================================
+
+.. automodule:: pyccea.initialization.continuous
+   :members:
+   :show-inheritance:
+   :undoc-members:
