+++ _sources/pages/pyccea.utils.models.rst.txt
@@ -1,0 +1,7 @@
+pyccea.utils.models module
+==========================
+
+.. automodule:: pyccea.utils.models
+   :members:
+   :show-inheritance:
+   :undoc-members:
