+++ _sources/pages/api.rst.txt
@@ -4,7 +4,7 @@ API Reference
 The following is the API reference for the PyCCEA package.
 
 .. toctree::
-   :maxdepth: 1
+   :maxdepth: 2
    :caption: Modules
 
    pyccea.coevolution
