+++ docs/source/pages/modules.rst
@@ -1,18 +1,0 @@
-API Reference
-=============
-
-The following is the API reference for the PyCCEA package.
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Modules
-
-   pages/pyccea.coevolution
-   pages/pyccea.cooperation
-   pages/pyccea.decomposition
-   pages/pyccea.evaluation
-   pages/pyccea.fitness
-   pages/pyccea.initialization
-   pages/pyccea.optimizers
-   pages/pyccea.projection
-   pages/pyccea.utils
