+++ _sources/pages/api.rst.txt
@@ -1,0 +1,18 @@
+API Reference
+=============
+
+The following is the API reference for the PyCCEA package.
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Modules
+
+   pyccea.coevolution
+   pyccea.cooperation
+   pyccea.decomposition
+   pyccea.evaluation
+   pyccea.fitness
+   pyccea.initialization
+   pyccea.optimizers
+   pyccea.projection
+   pyccea.utils
