### 2025-06-08 54f1d25
fix: keep init during module-level generation
