### 2025-06-08 a803e91
fix: generate module-level .rst files step
