+++ _sources/pages/pyccea.fitness.function.rst.txt
@@ -1,7 +1,0 @@
-pyccea.fitness.function module
-==============================
-
-.. automodule:: pyccea.fitness.function
-   :members:
-   :show-inheritance:
-   :undoc-members:
