+++ _sources/pages/pyccea.decomposition.dummy.rst.txt
@@ -1,7 +1,0 @@
-pyccea.decomposition.dummy module
-=================================
-
-.. automodule:: pyccea.decomposition.dummy
-   :members:
-   :show-inheritance:
-   :undoc-members:
