+++ _sources/pages/pyccea.coevolution.rst.txt
@@ -1,23 +1,85 @@
 pyccea.coevolution package
 ==========================
 
-.. automodule:: pyccea.coevolution
+Submodules
+----------
+
+pyccea.coevolution.ccea module
+------------------------------
+
+.. automodule:: pyccea.coevolution.ccea
    :members:
    :show-inheritance:
    :undoc-members:
 
-Submodules
-----------
+pyccea.coevolution.cceafs module
+--------------------------------
 
-.. toctree::
-   :maxdepth: 4
+.. automodule:: pyccea.coevolution.cceafs
+   :members:
+   :show-inheritance:
+   :undoc-members:
 
-   pyccea.coevolution.ccea
-   pyccea.coevolution.cceafs
-   pyccea.coevolution.ccfc
-   pyccea.coevolution.ccfsrfg1
-   pyccea.coevolution.ccfsrfg2
-   pyccea.coevolution.ccga
-   pyccea.coevolution.ccilfg
-   pyccea.coevolution.ccpstfg
-   pyccea.coevolution.ccsufg
+pyccea.coevolution.ccfc module
+------------------------------
+
+.. automodule:: pyccea.coevolution.ccfc
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyccea.coevolution.ccfsrfg1 module
+----------------------------------
+
+.. automodule:: pyccea.coevolution.ccfsrfg1
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyccea.coevolution.ccfsrfg2 module
+----------------------------------
+
+.. automodule:: pyccea.coevolution.ccfsrfg2
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyccea.coevolution.ccga module
+------------------------------
+
+.. automodule:: pyccea.coevolution.ccga
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyccea.coevolution.ccilfg module
+--------------------------------
+
+.. automodule:: pyccea.coevolution.ccilfg
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyccea.coevolution.ccpstfg module
+---------------------------------
+
+.. automodule:: pyccea.coevolution.ccpstfg
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyccea.coevolution.ccsufg module
+--------------------------------
+
+.. automodule:: pyccea.coevolution.ccsufg
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: pyccea.coevolution
+   :members:
+   :show-inheritance:
+   :undoc-members:
