### 2025-06-08 2aada4d
fix: force Sphinx to rebuild everything
