### 2025-06-09 b99297d
docs: improve api reference rst file
