### 2025-06-09 77040fc
fix: show module content
