+++ _sources/pages/api.rst.txt
@@ -4,8 +4,8 @@ API Reference
 The following is the API reference for the PyCCEA package.
 
 .. toctree::
-   :maxdepth: 2
-   :caption: Package Structure
+   :maxdepth: 1
+   :caption: Modules
 
    pyccea.coevolution
    pyccea.cooperation
@@ -16,61 +16,3 @@ The following is the API reference for the PyCCEA package.
    pyccea.optimizers
    pyccea.projection
    pyccea.utils
-
-
-Module Details
-==============
-
-.. automodule:: pyccea.coevolution
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
-
-.. automodule:: pyccea.cooperation
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
-
-.. automodule:: pyccea.decomposition
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
-
-.. automodule:: pyccea.evaluation
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
-
-.. automodule:: pyccea.fitness
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
-
-.. automodule:: pyccea.initialization
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
-
-.. automodule:: pyccea.optimizers
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
-
-.. automodule:: pyccea.projection
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
-
-.. automodule:: pyccea.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :imported-members:
