+++ _sources/pages/pyccea.optimizers.rst.txt
@@ -1,6 +1,11 @@
 pyccea.optimizers package
 =========================
 
+.. automodule:: pyccea.optimizers
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
 Submodules
 ----------
 
@@ -16,14 +21,6 @@ pyccea.optimizers.genetic\_algorithm module
 -------------------------------------------
 
 .. automodule:: pyccea.optimizers.genetic_algorithm
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: pyccea.optimizers
    :members:
    :show-inheritance:
    :undoc-members:
