### 2025-06-09 9c912d2
fix: generate stubs
