### 2025-06-09 3e159c2
feat: add a release workflow with changelog using poetry
