+++ .github/workflows/release.yml
@@ -1,0 +1,66 @@
+name: Release
+
+on:
+  push:
+    branches:
+      - main
+  workflow_dispatch:
+
+jobs:
+  release:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: Setup Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.10
+
+      - name: Install Poetry
+        run: |
+          curl -sSL https://install.python-poetry.org | python3 -
+          echo "$HOME/.local/bin" >> $GITHUB_PATH
+
+      - name: Install dependencies
+        run: poetry install --no-interaction --no-ansi
+
+      - name: Run tests
+        run: poetry run pytest
+
+      - name: "Check commit message and bump version (patch) if prefix 'release: ' is present"
+        id: bump_version
+        env:
+            GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+            LAST_COMMIT_MSG=$(git log -1 --pretty=%B)
+            echo "Last commit message: $LAST_COMMIT_MSG"
+            if [[ "$LAST_COMMIT_MSG" == release:* ]]; then
+            poetry version patch
+            git config user.name "github-actions"
+            git config user.email "github-actions@github.com"
+            git add pyproject.toml
+            git commit -m "ci: bump version [skip ci]"
+
+            git log $(git describe --tags --abbrev=0)..HEAD --pretty=format:"* %s (%an)" > CHANGELOG.md
+            git add CHANGELOG.md
+            git commit -m "docs: update changelog [skip ci]"
+
+            NEW_VERSION=$(poetry version -s)
+            git tag "v$NEW_VERSION"
+            git push origin main --tags
+            git push origin main
+            echo "version_bumped=true" >> $GITHUB_OUTPUT
+            else
+            echo "version_bumped=false" >> $GITHUB_OUTPUT
+            fi
+
+      - name: Publish to PyPI
+        if: steps.bump_version.outputs.version_bumped == 'true' && startsWith(github.ref, 'refs/tags/v')
+        env:
+          POETRY_PYPI_TOKEN_PYPI: ${{ secrets.PYPI_API_TOKEN }}
+        run: |
+          poetry config pypi-token.pypi $POETRY_PYPI_TOKEN_PYPI
+          poetry publish --no-interaction --build
