### 2025-06-09 3118ef1
fix: tool.poetry.include must be array
