### 2025-06-09 6f4107e
fix: python version in release workflow
