+++ .github/workflows/release.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v4
         with:
-          python-version: 3.10
+          python-version: "3.10.16"
 
       - name: Install Poetry
         run: |
