### 2025-06-09 1f50964
fix: pyproject config file
