+++ pyproject.toml
@@ -6,7 +6,6 @@ authors = ["Pedro Vinícius A. B. Venâncio <pedbrgs@gmail.com>"]
 license = "MIT"
 readme = "README.md"
 homepage = "https://github.com/pedbrgs/PyCCEA"
-python = ">=3.10"
 
 [tool.poetry.dependencies]
 python = ">=3.10"
