### 2025-06-09 0f7fc22
release: update README and test release workflow
