### 2025-06-09 22cd90a
fix: release workflow trigger
