+++ .github/workflows/release.yml
@@ -35,32 +35,38 @@ jobs:
         env:
             GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: |
+            LAST_TAG=$(git describe --tags --abbrev=0)
             LAST_COMMIT_MSG=$(git log -1 --pretty=%B)
             echo "Last commit message: $LAST_COMMIT_MSG"
             if [[ "$LAST_COMMIT_MSG" == release:* ]]; then
+            # Generate changelog BEFORE bump
+            git log $LAST_TAG..HEAD --pretty=format:"* %s (%an)" > CHANGELOG.md
+            git add CHANGELOG.md
+            git commit -m "docs: update changelog [skip ci]"
+
+            # Bump version
             poetry version patch
             git config user.name "github-actions"
             git config user.email "github-actions@github.com"
             git add pyproject.toml
             git commit -m "ci: bump version [skip ci]"
 
-            git log $(git describe --tags --abbrev=0)..HEAD --pretty=format:"* %s (%an)" > CHANGELOG.md
-            git add CHANGELOG.md
-            git commit -m "docs: update changelog [skip ci]"
-
             NEW_VERSION=$(poetry version -s)
             git tag "v$NEW_VERSION"
+
+            # Push all changes + tags
             git push origin main --tags
             git push origin main
+
             echo "version_bumped=true" >> $GITHUB_OUTPUT
             else
             echo "version_bumped=false" >> $GITHUB_OUTPUT
             fi
 
       - name: Publish to PyPI
-        if: steps.bump_version.outputs.version_bumped == 'true' && startsWith(github.ref, 'refs/tags/v')
+        if: steps.bump_version.outputs.version_bumped == 'true'
         env:
-          POETRY_PYPI_TOKEN_PYPI: ${{ secrets.PYPI_API_TOKEN }}
+            POETRY_PYPI_TOKEN_PYPI: ${{ secrets.PYPI_API_TOKEN }}
         run: |
-          poetry config pypi-token.pypi $POETRY_PYPI_TOKEN_PYPI
-          poetry publish --no-interaction --build
+            poetry config pypi-token.pypi $POETRY_PYPI_TOKEN_PYPI
+            poetry publish --no-interaction --build
