### 2025-06-09 ebce805
fix: release workflow trigger
