### 2025-06-09 6f696df
docs: update docs workflow trigger
