### 2025-06-09 df8eccf
release: update readme and test workflow triggers
