### 2025-06-09 667da06
fix: config user name first in release workflow
