### 2025-06-09 7b99676
release: update readme and test workflow triggers
