### 2025-06-09 c534c9a
ci: bump version [skip ci]
