+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyccea"
-version = "1.0.1"
+version = "1.0.2"
 description = "Cooperative co-evolutionary algorithms for feature selection in high-dimensional data"
 authors = ["Pedro Vinícius A. B. Venâncio <pedbrgs@gmail.com>"]
 license = "MIT"
