### 2025-06-09 e5b53f3
docs: update changelog [skip ci]
