+++ CHANGELOG.md
@@ -1,0 +1,1 @@
+* release: update readme and test workflow triggers (pedbrgs)
