### 2025-06-10 cc44bd6
docs: add new badges
