### 2025-06-10 d4ee8eb
docs: remove download per month badge
