+++ README.md
@@ -7,7 +7,6 @@
 [![status](https://joss.theoj.org/papers/dd1af373a617c14953d289c085c38d8f/status.svg)](https://joss.theoj.org/papers/dd1af373a617c14953d289c085c38d8f)
 ![License](https://img.shields.io/pypi/l/pyccea)
 ![Python Versions](https://img.shields.io/pypi/pyversions/pyccea)
-![PyPI - Downloads](https://img.shields.io/pypi/dm/pyccea)
 [![Downloads](https://pepy.tech/badge/pyccea)](https://pepy.tech/project/pyccea)
 
 
