### 2025-06-10 3f1e38b
chore: update docs workflow trigger
