### 2025-06-10 #8348 @pedbrgs
@maximtrp @geoffreyp Thank you for your initial suggestions. I have addressed all the points you raised.

> 1. [@pedbrgs](https://github.com/pedbrgs) Please correct me if I am wrong, but there is no documentation for your package now. If so, please create it (e.g. on [readthedocs.org](https://readthedocs.org)). You should document all public methods, provide a detailed tutorial and benchmark results (if any).

> 2. In general, the README is not sufficient as documentation.

Thank you for the recommendation. I have generated the documentation using Sphinx, encompassing installation instructions, tutorial, contribution guidelines, and citation information, along with an API reference detailing the definitions of classes, methods, and functions for each module of PyCCEA. The API reference section is generated automatically. I implemented a workflow to update the documentation automatically whenever a commit with the prefix `docs:` is made to the main branch.

> 3. [@pedbrgs](https://github.com/pedbrgs) It's a good practice to post release information on GitHub with a change log. So, I strongly suggest using a GitHub workflow for publishing your package on PyPi.

Nice suggestion! I also implemented a workflow to automatically perform releases whenever a commit with the prefix `release:` is made to the main branch. This workflow generates a `CHANGELOG.md` file containing the commits made since the last release.

> 4. [@pedbrgs](https://github.com/pedbrgs) Your package is currently using an outdated packaging config. Please update it to use [pyproject.toml](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/). There is even a [tutorial](https://packaging.python.org/en/latest/guides/modernize-setup-py-project/) for that.

Thank you for the heads-up! I made the change to use `pyproject.toml` in the Poetry format. It also significantly simplified the release workflow. Excellent suggestion!

> 5. Which version of Python should we use?

Good point! I hadn’t clearly specified that anywhere. I have now included the supported versions both in the installation instructions and as a badge to make it clear which Python versions are supported by the PyCCEA package.

> 6. How can we contribute to the project? The CONTRIB file is missing ([The github documentation](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)).

I have added the contribution guidelines! They are included both in the `CONTRIBUTING.md` file at the root of the repository and in the package documentation.

> 7. I was unable to test the example from your README by following your instructions (pip install pyccea).

My apologies for that, it was an oversight on my part. It turns out that switching to `pyproject.toml` configuration helped me resolve this issue. In any case, I believe everything is working properly now. I tested it in a fresh `pyenv` environment as well as in a Docker container.

Thanks once again! With all these suggestions, the package has already seen a significant improvement in terms of documentation and automation.
