### 2025-06-10 #8348 @geoffreyp
Great update!
Here are some more of my remarks on the documentation and installation:

**Installation:**
Great addition of pyproject.toml!
For the installation, I only had an issue with Python 3.11 due to pyarrow (ERROR: Could not build wheels for pyarrow, which is required to install pyproject.toml-based projects), but it works with 3.12.
In your CI/CD with your tests, [I see that you only test for 3.10](https://github.com/pedbrgs/PyCCEA/blob/main/.github/workflows/tests.yml); you should test all versions that you consider compatible with PyCCEA.

**For the documentation:**
For auto-generated documentation, it's not advisable to include only the title of the algorithm or article. It would be more practical to add a small paragraph explaining the algorithm without having to read the entire paper.

Given the modular architecture presented in your paper, I expect to see in the documentation:
- pyccea.coevolution.decomposition
- pyccea.coevolution.cooperation
- pyccea.coevolution.fitness
- pyccea.coevolution.evaluation
- pyccea.coevolution.optimizers
- pyccea.benchmark
- ...

So, you should add them (I see that they are in the code with good documentation).

You could also make the documentation clearer with a "User Guide" section by globally describing each package with a redirection to the auto-generated technical description (example: [MOEAD Framework User Guide](https://moead-framework.github.io/framework/html/user_guide.html) and [MOEAD Framework API](https://moead-framework.github.io/framework/html/api.html)).

Similarly, to help users envision using PyCCEA, you can provide examples of customization, such as how to implement an optimizer or a new fitness function for use in PyCCEA.
