### 2025-06-13 8896543
chore: remove pytest install redundancy from test workflow
