### 2025-06-13 760db53
fix: fold normalization
