### 2025-06-13 ebd3a1a
fix: fold normalization
