### 2025-06-13 b5dc56e
fix: remove future no silent downcasting
