### 2025-06-18 6b48f94
feat: store tuning time as an attribute for CCPSTFG
