### 2025-06-19 576516a
chore: test across multiple Python versions
