### 2025-06-19 eb74bb7
chore: remove requirements.txt since PyCCEA uses poetry directly
