+++ requirements.txt
@@ -1,16 +1,0 @@
-imbalanced-learn==0.13.0
-kneed==0.8.5
-numpy==1.24.4
-pandas==2.0.3
-scikit-learn==1.6.1
-scipy==1.10.1
-toml==0.10.2
-tqdm==4.65.0
-tqdm-joblib==0.0.4
-pyarrow==17.0.0
-pytest==8.3.5
-pytest-cov==6.1.1
-pytest-asyncio==0.26.0
-pytest-deadfixtures==2.2.1
-pytest-mock==3.14.0
-fsspec==2025.5.1
