### 2025-06-22 13c1aa7
docs: add a complete user guide
