+++ docs/source/pages/tutorials.rst
@@ -1,0 +1,12 @@
+Tutorials
+====================
+
+Here are some tutorials on how to use and customize cooperative co-evolutionary algorithms for
+feature selection using the PyCCEA package.
+
+.. toctree::
+   :maxdepth: 1
+   :numbered: 2
+   :caption: Contents
+
+   examples/quickstart
