### 2025-06-22 fd0dc56
docs: remove additional README files
