+++ pyccea/decomposition/README.md
@@ -1,12 +1,0 @@
-## Available feature grouping
-
-The following feature grouping strategies are currently implemented in the PyCCEA package:
-
-| Strategy | Reference |
-|------|--------|
-|  Sequential   |   [Rashid, A. N. M., et al. "A novel penalty-based wrapper objective function for feature selection in Big Data using cooperative co-evolution." IEEE Access 8 (2020): 150113-150129.](https://ieeexplore.ieee.org/abstract/document/9167222/)     |
-|  Random    |  [Rashid, A. N. M., et al. "Cooperative co-evolution for feature selection in Big Data with random feature grouping." Journal of Big Data 7.1 (2020): 1-42.](https://journalofbigdata.springeropen.com/articles/10.1186/s40537-020-00381-y)      |
-|  Symmetric Uncertainty     |  [Song, Xian-Fang, et al. "Variable-size cooperative coevolutionary particle swarm optimization for feature selection on high-dimensional data." IEEE Transactions on Evolutionary Computation 24.5 (2020): 882-895.](https://ieeexplore.ieee.org/abstract/document/8966273?casa_token=cjoOD2JFHqsAAAAA:ntjT3yV20eoBU8I8n-lSg3Xtj_p4zc8WfW_HWKe5eL_KhQSRLF9S-JQwLRWapA8y2URR1zd97Q)      |
-|  Clustering    |  [Li, H., et al. "MLFS-CCDE: multi-objective large-scale feature selection by cooperative coevolutionary differential evolution." Memetic Computing 13 (2021): 1-18.](https://link.springer.com/article/10.1007/s12293-021-00328-7)      |
-|  Interaction Learning    |  [Hou, Yaqing, et al. "A correlation-guided cooperative coevolutionary method for feature selection via interaction learning-based space division." Swarm and Evolutionary Computation 93 (2025): 101846.](https://www.sciencedirect.com/science/article/pii/S2210650225000045?casa_token=W4_aOJh1xx4AAAAA:YyrtjVSunL1cDbLneSXpEegOfVTjUrnMBzChgq23SM8Dk4Vasxjaa_Fjoen4zkG4FXGBALxbtQ)      |
-|  Projection-based self-tuning     |  [Venâncio, P. V. A. B., et al. "A self-tuning decomposition strategy in cooperative co-evolutionary algorithms for high-dimensional feature selection." Knowledge-Based Systems (2025): 113327.](https://www.sciencedirect.com/science/article/pii/S0950705125003740)      |
