+++ _sources/index.rst.txt
@@ -1,19 +1,19 @@
 Documentation
-====================
+=============
 
-PyCCEA is an open-source Python package for feature selection in high-dimensional datasets using
-cooperative co-evolutionary algorithms (CCEAs). It provides a modular framework that enables
+**PyCCEA** is an open-source Python package for feature selection in high-dimensional datasets using
+**cooperative co-evolutionary algorithms** (CCEAs). It provides a modular framework that enables
 researchers and practitioners to implement, evaluate, and extend CCEAs through interchangeable
-components such as decomposition strategies, cooperation schemes, fitness evaluation methods, and
-evolutionary optimizers.
+components such as decomposition strategies, collaboration techniques, fitness evaluation methods,
+and evolutionary optimizers.
 
 .. toctree::
    :maxdepth: 1
-   :numbered: 4
    :caption: Contents
 
    pages/installation
-   pages/quickstart
+   pages/user_guide
+   pages/tutorials
    pages/api
    pages/contributing
    pages/citation
