+++ _sources/pages/user_guide.rst.txt
@@ -1,0 +1,650 @@
+User Guide
+==========
+
+This user guide provides a comprehensive overview of how to use **PyCCEA**, a Python package for
+feature selection based on **Cooperative Co-Evolutionary Algorithms (CCEAs)**. It covers both the
+theoretical foundations and practical aspects of the framework, including configuration files,
+algorithm components, benchmark and result reproducibility.
+
+Whether you're a researcher aiming to benchmark feature selection algorithms or a practitioner
+optimizing high-dimensional datasets, this guide will walk you through the key concepts, options,
+and implementation details necessary to use PyCCEA effectively.
+
+Feature selection problem
+-------------------------
+
+Given an input matrix :math:`\mathbf{X} \in \mathbb{R}^{m \times n}`, representing a dataset with
+:math:`m` instances and :math:`n` features, and an output vector
+:math:`\mathbf{y} \in \mathbb{B}^{m \times 1}` containing target labels for each instance, the
+feature selection problem can be formulated as a **combinatorial optimization task**.
+
+The goal is to identify a subset of features :math:`S^* \subseteq \mathcal{F}` that maximizes a
+predefined objective function :math:`f`, where :math:`\mathcal{F} = \{f_1, f_2, \dots, f_n\}`
+denotes the full feature set.
+
+The objective function :math:`f` evaluates the quality or relevance of selected features,
+typically by measuring the performance of a learning model trained and tested on the selected
+feature subset. The function :math:`f` may consider factors such as predictive accuracy,
+model complexity, interpretability, computational efficiency, or domain-specific criteria.
+Additional constraints or regularization terms may also be included to control aspects like the
+number of selected features or redundancy among them.
+
+An arbitrary solution :math:`S` corresponds to a subset of selected feature indices, such that
+:math:`\|S\| = n_s \leq n`. For instance, if :math:`S = \{i, j, k\}`, it implies that the
+:math:`i`-th, :math:`j`-th, and :math:`k`-th features are selected from the original feature set,
+resulting in a data subset :math:`\mathbf{X}[:, S] = \mathbf{X}^{'} \in \mathbb{R}^{m \times n_s}`.
+
+Formally, the feature selection problem is defined as:
+
+.. math::
+
+    S^{*} = \operatorname*{arg\,max}_{S \subseteq \{f_1, f_2, \dots, f_n\}} f(\mathbf{X}^{'}, \mathbf{y})
+
+where :math:`f(\mathbf{X}^{'}, \mathbf{y})` represents the fitness value of the selected feature
+subset :math:`\mathbf{X}^{'} = \mathbf{X}[:, S]`.
+
+Cooperative co-evolutionary algorithms
+--------------------------------------
+
+Cooperative co-evolutionary algorithms (CCEAs) are a class of evolutionary computation techniques
+designed to tackle complex, large-scale optimization problems by decomposing them into smaller,
+more manageable subproblems. Each subproblem is addressed by an independent subpopulation, which
+is evolved by its own dedicated evolutionary algorithm. After an optimization cycle, the
+evaluation of each individual is determined based on their effectiveness in collaborating with
+representative individuals from other subpopulations to collectively compose one or more complete
+solutions (context vectors). This iterative process continues until convergence or a termination
+condition is met, as shown in :numref:`CCEA_framework_figure`.
+
+.. _CCEA_framework_figure:
+.. figure:: ../../figures/CCEA.png
+   :align: center
+   :width: 60%
+   :alt: CCEA Framework
+
+   A general framework of a Cooperative Co-Evolutionary Algorithm (CCEA).
+
+Problem decomposition
+^^^^^^^^^^^^^^^^^^^^^
+
+In feature selection, problem decomposition, also referred to as feature grouping in the
[…]
