### 2025-06-22 a841a8c
docs: improve tutorials
