+++ docs/source/pages/tutorials.rst
@@ -1,12 +1,322 @@
 Tutorials
-====================
+=========
 
-Here are some tutorials on how to use and customize cooperative co-evolutionary algorithms for
-feature selection using the PyCCEA package.
+This section provides step-by-step tutorials on how to use and customize cooperative
+co-evolutionary algorithms for feature selection using the PyCCEA package.
 
-.. toctree::
-   :maxdepth: 1
-   :numbered: 2
-   :caption: Contents
+.. contents::
+   :local:
+   :depth: 2
+   :backlinks: entry
 
-   examples/quickstart
+How to use a baseline CCEA?
+---------------------------
+
+This tutorial demonstrates how to use the ``CCFSRFG1`` algorithm — a cooperative co-evolutionary
+algorithm (CCEA) variant with random feature grouping — to perform feature selection on the
+Wisconsin Diagnostic Breast Cancer (WDBC) dataset.
+
+In this example, you will:
+
+- Load the dataset using the ``DataLoader``.
+- Load the dataset and algorithm configuration files.
+- Run the optimization process.
+
+We start by importing the necessary modules and classes:
+
+.. code-block:: python
+
+    import toml
+    import importlib.resources
+    from pyccea.coevolution import CCFSRFG1
+    from pyccea.utils.datasets import DataLoader
+
+- ``toml`` is used to parse configuration files.
+- ``importlib.resources`` helps access files inside a package.
+- ``CCFSRFG1`` is the cooperative co-evolution algorithm you'll run.
+- ``DataLoader`` is a utility to prepare the dataset.
+
+Next, we load the configuration for the dataset from a `.toml` file:
+
+.. code-block:: python
+
+    with importlib.resources.open_text("pyccea.parameters", "dataloader.toml") as toml_file:
+        data_conf = toml.load(toml_file)
+
+This code looks for the file ``dataloader.toml`` inside the ``pyccea.parameters`` package and loads
+its content into a dictionary.
+
+We then create and prepare the dataset using the ``DataLoader``:
+
+.. code-block:: python
+
+    dataloader = DataLoader(dataset="wdbc", conf=data_conf)
+    dataloader.get_ready()
+
+- ``dataset="wdbc"`` specifies the dataset to use (WDBC in this case).
+- ``conf=data_conf`` passes the parameters we just loaded.
+- ``get_ready()`` performs any necessary preprocessing and splits the data.
+
+We now load the configuration for the CCFSRFG1 algorithm:
+
+.. code-block:: python
+
+    with importlib.resources.open_text("pyccea.parameters", "ccfsrfg.toml") as toml_file:
+        ccea_conf = toml.load(toml_file)
+
+This step reads the ``ccfsrfg.toml`` file and loads the algorithm's hyperparameters into a
+dictionary.
+
+Now we're ready to run the algorithm:
+
+.. code-block:: python
+
+    ccea = CCFSRFG1(data=dataloader, conf=ccea_conf, verbose=False)
+    ccea.optimize()
+
+- ``CCFSRFG1(...)`` initializes the algorithm with data and configuration.
+- ``optimize()`` starts the evolutionary process for feature selection.
+
+Once the optimization is complete, the best feature subset is stored in the ``best_context_vector``
+attribute of the ``CCFSRFG1`` object, a binary vector where:
+
+  - `1` indicates that a feature is selected.
+  - `0` indicates that a feature is excluded.
+
+Some CCEAs, like ``CCFSRFG1``, may reorder the features during the decomposition phase. In these
+cases, you must map the selected features back to their original names using the reordered
+indices.
+
+.. code-block:: python
+
+    # Get original feature column names
+    feature_cols = dataloader.data.columns
[…]
