### 2025-06-22 #8348 @pedbrgs
@geoffreyp Thank you for your considerations, and sorry for taking a bit longer to go through everything. There were quite a few points to consider, and I needed some extra time to address them properly during my free time.

> Great addition of pyproject.toml!
For the installation, I only had an issue with Python 3.11 due to pyarrow (ERROR: Could not build wheels for pyarrow, which is required to install pyproject.toml-based projects), but it works with 3.12.
In your CI/CD with your tests, [I see that you only test for 3.10](https://github.com/pedbrgs/PyCCEA/blob/main/.github/workflows/tests.yml); you should test all versions that you consider compatible with PyCCEA.

Could you please provide more details about your PyArrow error? The version I specified in `pyproject.toml` is 20.0.0, which, according to the documentation, is currently compatible with Python 3.9, 3.10, 3.11, 3.12, and 3.13 ([Installing PyArrow](https://arrow.apache.org/docs/python/install.html#python-compability)).
Additionally, I've already added tests for all Python versions between 3.10 and 3.13, and they all passed successfully in the GitHub workflow. I also performed a local test by creating a Conda environment with Python 3.11 and installing the current version of PyCCEA (1.0.2), and I didn't encounter any issues. PyArrow worked well.

> You could also make the documentation clearer with a "User Guide" section by globally describing each package with a redirection to the auto-generated technical description (example: [MOEAD Framework User Guide](https://moead-framework.github.io/framework/html/user_guide.html) and [MOEAD Framework API](https://moead-framework.github.io/framework/html/api.html)).

Thank you for the suggestion. I have already developed a comprehensive user guide that provides both theoretical and practical insights. It explains how the feature selection problem is formulated as a combinatorial optimization task and offers a high-level overview of cooperative co-evolutionary algorithms and their core components. The guide also outlines the available implementation strategies for each CCEA component. Additionally, it includes dedicated sections describing the configuration parameters for both the data loader and the algorithm. A benchmarking section is provided as well, detailing the classification and regression datasets currently included in the package, along with their properties, and listing and describing the implemented CCEA variants available for experimentation. Practical guidelines to support reproducibility are also included. As you suggested, throughout the user guide I included several references to the API Reference to help users navigate between conceptual explanations and technical usage.

> Similarly, to help users envision using PyCCEA, you can provide examples of customization, such as how to implement an optimizer or a new fitness function for use in PyCCEA.

Excellent suggestion! I’ve improved the Tutorials section by splitting it into two parts: "How to use a baseline CCEA" and "How to customize your a CCEA", each with clear, step-by-step examples to guide users through the process.
