### 2025-06-26 bce8862
paper: add reference to the documentation in the article
