+++ paper/paper.md
@@ -37,7 +37,7 @@ The package currently adopts a wrapper-based evaluation approach, where subsets 
 
 In addition to its evolutionary components, `PyCCEA` integrates machine learning and data processing functionalities built on top of widely adopted libraries, including `scikit-learn` [@Pedregosa:2011] and `Pandas` [@McKinney:2010]. In view of this, users can take advantage of built-in support for models (e.g., KNN, SVM, Random Forest), metrics (e.g., accuracy, $F_1$-score), and data operations (e.g., normalization, cross-validation, splitting, preprocessing). This tight integration ensures compatibility with existing machine learning workflows while maintaining flexibility and ease of use.
 
-To further support empirical studies, `PyCCEA` includes a benchmarking module composed of well-known datasets (e.g., DLBCL, lung cancer) [@Kelly:2017] and baseline CCEAs (e.g., CCEAFS [@Rashid:2020a], CCFSRFG1 [@Rashid:2020b]). This enables standardized evaluation and comparison of new strategies in a reproducible and transparent manner. To the best of our knowledge, `PyCCEA` is the first open-source package specifically designed for cooperative co-evolution in the context of feature selection, providing a solid foundation for research and practical application in high-dimensional machine learning problems.
+To further support empirical studies, `PyCCEA` includes a benchmarking module composed of well-known datasets (e.g., DLBCL, lung cancer) [@Kelly:2017] and baseline CCEAs (e.g., CCEAFS [@Rashid:2020a], CCFSRFG1 [@Rashid:2020b]). This enables standardized evaluation and comparison of new strategies in a reproducible and transparent manner. To the best of our knowledge, `PyCCEA` is the first open-source package specifically designed for cooperative co-evolution in the context of feature selection, providing a solid foundation for research and practical application in high-dimensional machine learning problems. For comprehensive details on installation, usage, and API, please refer to the official [PyCCEA documentation](https://pedbrgs.github.io/PyCCEA/).
 
 ![An overview of the `PyCCEA` package, its modules and underlying Python dependencies. \label{fig:architecture}](figures/architecture.png)
 
