### 2025-06-28 3c84bb3
paper: include references to other feature selection packages
