+++ paper/paper.md
@@ -35,9 +35,9 @@ Cooperative co-evolutionary algorithms (CCEAs) offer a promising approach for ta
 
 The package currently adopts a wrapper-based evaluation approach, where subsets of features are assessed using machine learning models to guide the search process. However, this design choice is not restrictive. `PyCCEA` has been structured to accommodate future extensions supporting other evaluation paradigms, such as filter-based, embedded, or hybrid methods. These alternatives are under consideration for upcoming releases to broaden the range of applicable scenarios and research directions.
 
-In addition to its evolutionary components, `PyCCEA` integrates machine learning and data processing functionalities built on top of widely adopted libraries, including `scikit-learn` [@Pedregosa:2011] and `Pandas` [@McKinney:2010]. In view of this, users can take advantage of built-in support for models (e.g., KNN, SVM, Random Forest), metrics (e.g., accuracy, $F_1$-score), and data operations (e.g., normalization, cross-validation, splitting, preprocessing). This tight integration ensures compatibility with existing machine learning workflows while maintaining flexibility and ease of use.
+In addition to its evolutionary components, `PyCCEA` integrates machine learning and data processing functionalities built on top of widely adopted libraries, including `scikit-learn` [@Pedregosa:2011] and `Pandas` [@McKinney:2010]. In view of this, users can take advantage of built-in support for models (e.g., KNN, SVM, Random Forest), metrics (e.g., accuracy, $F_1$-score), and data operations (e.g., normalization, cross-validation, splitting, preprocessing). This tight integration ensures compatibility with existing machine learning workflows while maintaining flexibility and ease of use. For comprehensive details on installation, usage, and API, please refer to the official [PyCCEA documentation](https://pedbrgs.github.io/PyCCEA/).
 
-To further support empirical studies, `PyCCEA` includes a benchmarking module composed of well-known datasets (e.g., DLBCL, lung cancer) [@Kelly:2017] and baseline CCEAs (e.g., CCEAFS [@Rashid:2020a], CCFSRFG1 [@Rashid:2020b]). This enables standardized evaluation and comparison of new strategies in a reproducible and transparent manner. To the best of our knowledge, `PyCCEA` is the first open-source package specifically designed for cooperative co-evolution in the context of feature selection, providing a solid foundation for research and practical application in high-dimensional machine learning problems. For comprehensive details on installation, usage, and API, please refer to the official [PyCCEA documentation](https://pedbrgs.github.io/PyCCEA/).
+To further support empirical studies, `PyCCEA` includes a benchmarking module composed of well-known datasets (e.g., DLBCL, lung cancer) [@Kelly:2017] and baseline CCEAs (e.g., CCEAFS [@Rashid:2020a], CCFSRFG1 [@Rashid:2020b]). This enables standardized evaluation and comparison of new strategies in a reproducible and transparent manner. To the best of our knowledge, `PyCCEA` is the first open-source package specifically designed for cooperative co-evolution in the context of feature selection, providing a solid foundation for research and practical application in high-dimensional machine learning problems. In contrast, other popular packages such as `scikit-learn` [@Pedregosa:2011], `MLXtend` [@Raschka:2018], and `Feature-engine` [@Galli:2021] rely on more traditional approaches, including statistical filters (e.g., ANOVA, correlation, mutual information), model-based embedded methods (e.g., tree importance, model coefficients), and deterministic or greedy searches (e.g., sequential or exhaustive search, recursive elimination). These methods, while effective in low to moderate dimensions, often struggle with scalability and performance in high-dimensional spaces, where complex feature interactions and search space combinatorics demand more flexible and exploratory algorithms such as CCEAs.
 
 ![An overview of the `PyCCEA` package, its modules and underlying Python dependencies. \label{fig:architecture}](figures/architecture.png)
 
