+++ paper/paper.bib
@@ -17,7 +17,7 @@
     journal={Journal of Machine Learning Research},
     volume={12},
     pages={2825--2830},
-    url={https://dl.acm.org/doi/10.5555/1953048.2078195},
+    url={10.5555/1953048.2078195},
     year={2011}
 }
 
@@ -32,7 +32,8 @@
 
 @article{Ma:2018,
     title={A survey on cooperative co-evolutionary algorithms},
-    author={Ma, Xiaoliang and Li, Xiaodong and Zhang, Qingfu and Tang, Ke and Liang, Zhengping and Xie, Weixin and Zhu, Zexuan},
+    author={Ma, Xiaoliang and Li, Xiaodong and Zhang, Qingfu and Tang, Ke and Liang, Zhengping and
+    Xie, Weixin and Zhu, Zexuan},
     journal={IEEE Transactions on Evolutionary Computation},
     volume={23},
     number={3},
@@ -42,8 +43,22 @@
     publisher={IEEE}
 }
 
+@article{Raschka:2018,
+    title={MLxtend: Providing machine learning and data science utilities and extensions to
+    Python's scientific computing stack},
+    author={Sebastian Raschka},
+    journal={The Journal of Open Source Software},
+    volume={3},
+    number={24},
+    pages={638},
+    year={2018},
+    doi={10.21105/joss.00638},
+    publisher={The Open Journal}
+}
+
 @article{Song:2020,
-    title={Variable-size cooperative coevolutionary particle swarm optimization for feature selection on high-dimensional data},
+    title={Variable-size cooperative coevolutionary particle swarm optimization for feature
+    selection on high-dimensional data},
     author={Song, Xian-Fang and Zhang, Yong and Guo, Yi-Nan and Sun, Xiao-Yan and Wang, Yong-Li},
     journal={IEEE Transactions on Evolutionary Computation},
     volume={24},
@@ -55,7 +70,8 @@
 }
 
 @article{Rashid:2020a,
-    title={A novel penalty-based wrapper objective function for feature selection in Big Data using cooperative co-evolution},
+    title={A novel penalty-based wrapper objective function for feature selection in Big Data 
+    using cooperative co-evolution},
     author={Rashid, ANM Bazlur and Ahmed, Mohiuddin and Sikos, Leslie F and Haskell-Dowland, Paul},
     journal={IEEE Access},
     volume={8},
@@ -77,8 +93,21 @@
     publisher={SpringerOpen}
 }
 
+@article{Galli:2021,
+    title={Feature-engine: A Python package for feature engineering for machine learning},
+    author={Soledad Galli},
+    journal={The Journal of Open Source Software},
+    volume={6},
+    number={65},
+    pages={3642},
+    year={2021},
+    doi={10.21105/joss.03642},
+    publisher={The Open Journal}
+}
+
 @article{Firouznia:2023,
-    title={Adaptive cooperative coevolutionary differential evolution for parallel feature selection in high-dimensional datasets},
+    title={Adaptive cooperative coevolutionary differential evolution for parallel feature 
+    selection in high-dimensional datasets},
     author={Firouznia, Marjan and Ruiu, Pietro and Trunfio, Giuseppe A},
     journal={The Journal of Supercomputing},
     pages={1--30},
@@ -89,7 +118,8 @@
 
 @article{Zhou:2024,
     title={A Knowledge-Guided Competitive Co-Evolutionary Algorithm for Feature Selection},
-    author={Zhou, Junyi and Zheng, Haowen and Li, Shaole and Hao, Qiancheng and Zhang, Haoyang and Gao, Wenze and Wang, Xianpeng},
+    author={Zhou, Junyi and Zheng, Haowen and Li, Shaole and Hao, Qiancheng and Zhang, Haoyang and 
[…]
