### 2025-06-28 9a836e7
paper: fix scikit-learn reference
