### 2025-06-28 d301224
paper: fix scikit-learn reference
