### 2025-06-28 e58e94b
paper: replace url field by doi in scikit-learn reference
