+++ paper/paper.bib
@@ -17,7 +17,7 @@
     journal={Journal of Machine Learning Research},
     volume={12},
     pages={2825--2830},
-    url={10.48550/arXiv.1201.0490},
+    doi={10.48550/arXiv.1201.0490},
     year={2011}
 }
 
