### 2025-06-28 #8348 @pedbrgs
Hi @geoffreyp, thank you for the suggestions regarding the paper!

> 1. You can add references to the documentation in the article.

Done! The paper includes a reference to the documentation.

> 2. It would be good to include references to other software (with links to the article and the code) for feature selection, even if it is not cooperative, to emphasize the difference from existing solutions.

Also addressed! The paper compares PyCCEA with three widely used feature selection packages (scikit-learn, MLXtend, and Feature-engine), highlighting that they rely on traditional search and filter methods rather than cooperative co-evolution. The last two are also published in JOSS, which further supports the relevance of the comparison.

I hope these additions meet your expectations. Please let me know if you have any further suggestions or feedback.
