### 2025-06-28 b653ede
chore: clear memory in the CCPSTFG
