### 2025-07-10 #8348 @maximtrp
@pedbrgs I went through the docs, installed and tested your package — everything's working fine now. Just curious about one thing: why are the dependency requirements so strict? When you pin everything to specific versions, it makes it harder for users to work with newer packages in their environments. They either can't upgrade their dependencies or have to force override yours, which isn't good. Please consider loosening up the version constraints a bit.

For example,
```
numpy = ">=2.2.0,<3.0.0"
```
