### 2025-07-11 fd2e7d8
chore: relax version constraints in pyproject.toml to improve compatibility
