+++ pyproject.toml
@@ -11,10 +11,10 @@ homepage = "https://github.com/pedbrgs/PyCCEA"
 python = ">=3.10,<3.14"
 imbalanced-learn = "0.13.0"
 kneed = "0.8.5"
-numpy = "2.2.0"
-pandas = "2.2.3"
-scikit-learn = "1.6.1"
-scipy = "1.15.3"
+numpy = ">=2.2,<3.0"
+pandas = ">=2.2,<3.0"
+scikit-learn = ">=1.6,<2.0"
+scipy = ">=1.15,<2.0"
 toml = "0.10.2"
 tqdm = "4.65.0"
 tqdm-joblib = "0.0.4"
