### 2025-07-11 #8348 @pedbrgs
Thanks for the suggestion, @maximtrp! That's a great point.

I've just updated the dependency constraints in `pyproject.toml` to be more permissive, allowing minor and patch version upgrades while staying within the same major versions. This should make it easier to integrate the package into environments with newer versions of common libraries like `numpy`, `pandas`, and `scikit-learn`.

I noticed that you’ve completed the review checklist too! Thank you very much for your thorough feedback and for contributing to making PyCCEA a more robust and user-friendly package.
