### 2025-08-02 2f28557
fix: DE docstring
