### 2025-08-02 a8b298e
feat: add linear and nonlinear synthetic datasets for binary classification task
