### 2025-08-02 b7af589
fix: paths for linear and non-linear datasets
