### 2025-08-10 #8348 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
