### 2025-08-11 #8348 @crvernon
:wave: @pedbrgs - this submission is looking good.  I just need you to address the following before I accept:

In your paper:

- [ ] LINE 90:  The "p" in "python" should be capitalized.  You can do this in your bib file by placing the characters you wish to maintain formatting of in curly brackets.

Let me know when this has been done and I'll accept.  Thank you.
