### 2025-08-11 9a25e1a
paper: capitalize Python in the references
