+++ paper/paper.bib
@@ -44,8 +44,8 @@
 }
 
 @article{Raschka:2018,
-    title={MLxtend: Providing machine learning and data science utilities and extensions to
-    Python's scientific computing stack},
+    title={{MLxtend: Providing machine learning and data science utilities and extensions to
+    Python's scientific computing stack}},
     author={Sebastian Raschka},
     journal={The Journal of Open Source Software},
     volume={3},
@@ -94,7 +94,7 @@
 }
 
 @article{Galli:2021,
-    title={Feature-engine: A Python package for feature engineering for machine learning},
+    title={{Feature-engine: A Python package for feature engineering for machine learning}},
     author={Soledad Galli},
     journal={The Journal of Open Source Software},
     volume={6},
