### 2025-08-11 #8348 @pedbrgs
@crvernon Thanks for the feedback!
I have updated the paper so that "Python" is correctly capitalized in the bibliography as requested.
